% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap}
\alias{block_bootstrap}
\title{Apply block-bootstrapping procedure}
\usage{
block_bootstrap(incidence_input, block_size = 10, days_incl = 21)
}
\arguments{
\item{days_incl}{}
}
\value{

}
\description{
Apply block-bootstrapping procedure
}
