% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap_overlap_func}
\alias{block_bootstrap_overlap_func}
\title{Title}
\usage{
block_bootstrap_overlap_func(
  incidence_vector,
  block_size = 10,
  keep_weekdays_aligned = TRUE
)
}
\arguments{
\item{keep_weekdays_aligned}{}
}
\value{

}
\description{
Title
}
