% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{bootstrap_params}
\alias{bootstrap_params}
\title{Bootstrapping parameters}
\arguments{
\item{N_bootstrap_replicates}{integer. Number of bootstrap samples.}

\item{combine_bootstrap_and_estimation_uncertainties}{boolean.
If TRUE, the uncertainty interval reported is the union of
the highest posterior density interval from the Re estimation
with the confidence interval from the boostrapping of time series
of observations.}
}
\description{
Bootstrapping parameters
}
