% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{bootstrap_return}
\alias{bootstrap_return}
\title{Bootstrapping pipe}
\value{
Effective reproductive estimates through time with confidence interval boundaries.
If \code{output_Re_only} is \code{FALSE}, then transformations made
on the input observations during calculations are output as well.
}
\description{
Bootstrapping pipe
}
