% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{build_delay_distribution}
\alias{build_delay_distribution}
\title{Build a delay distribution vector}
\usage{
build_delay_distribution(
  parm1,
  parm2,
  distribution_type = "gamma",
  max_quantile = 0.999,
  offset_by_one = FALSE
)
}
\arguments{
\item{parm1}{numeric. If \code{distribution_type=="gamma"}, \code{parm1} is the shape parameter.}

\item{parm2}{numeric. If \code{distribution_type=="gamma"}, \code{parm2} is the scale parameter.}

\item{distribution_type}{string. Options are "gamma".}

\item{max_quantile}{numeric value between 0 and 1. TODO write what max_quantile does}

\item{offset_by_one}{boolean. Gamma distribution comes from fit on (raw_data + 1) to accommodate zeroes in the raw_data.}
}
\value{
numeric vector.
}
\description{
Only allows for gamma distributions for now.
}
\examples{
#TODO add example
}
