% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH_linelist}
\alias{CH_linelist}
\title{Linelist of COVID-19 patients in Switzerland}
\format{
A data frame with 32'052 rows and 4 columns:
\describe{
\item{confirmation_date}{date of case confirmation in YYYY-mm-dd format}
\item{symptom_onset_date}{If applicable and reported, date of onset of symptoms in YYYY-mm-dd format}
\item{hospital_admission_date}{If applicable, date of admission to a hospital in YYYY-mm-dd format}
\item{death_date}{If applicable, date of death of the patient in YYYY-mm-dd format}
}
}
\source{
Federal Office of Public Health
}
\usage{
CH_linelist
}
\description{
A dataset containing a linelist of Swiss COVID-19 patients.
Each row refers to a particular patient.
The 'confirmation_date' column refers to the date at which a positive COVID-19 test is reported.
If applicable, a date of hospital admission is included.
If applicable, a date of death is included.
If applicable and known, a date of onset of symptoms is included.
}
\details{
This linelist is provided by the Federal Office of Public Health in Switzerland.
}
\keyword{datasets}
