% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH_simulated_linelist}
\alias{CH_simulated_linelist}
\title{Synthetic linelist of COVID-19 patients}
\format{
A data frame with 31'950 rows and 2 columns:
\describe{
\item{confirmation_date}{date of case confirmation in YYYY-mm-dd format}
\item{symptom_onset_date}{If simulated, date of onset of symptoms in YYYY-mm-dd format}
}
}
\usage{
CH_simulated_linelist
}
\description{
A synthetic dataset containing a simulated linelist of Swiss COVID-19 patients from February to June 2020.
The incidence that can be derived from aggregating this synthetic linelist
corresponds to or closely matches the
incidence numbers reported for Switzerland by the Federal Office of Public Health.
}
\details{
In this simulated dataset, each row refers to a particular patient.
The 'confirmation_date' column refers to the date at which a positive COVID-19 test is reported.
For each patient, with a probability < 1, a date of onset of symptoms was drawn from a probability distribution.
Otherwise, the date of onset of symptoms was left as NA.
}
\keyword{datasets}
