% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{combine_incubation_with_reporting_delay}
\alias{combine_incubation_with_reporting_delay}
\title{Build a waiting time distribution from the convolution of two gamma distributions}
\usage{
combine_incubation_with_reporting_delay(
  parm1_incubation,
  parm2_incubation,
  parm1_onset_to_report,
  parm2_onset_to_report,
  distribution_type_incubation = "gamma",
  distribution_type_onset_to_report = "gamma",
  max_quantile = 0.9999,
  incubation_offset_by_one = FALSE,
  onset_to_report_offset_by_one = FALSE
)
}
\arguments{
\item{parm1_incubation}{numeric. Shape parameter of the gamma distribution representing the delay between infection and symptom onset.}

\item{parm2_incubation}{numeric. Scale parameter of the gamma distribution representing the delay between infection and symptom onset.}

\item{parm1_onset_to_report}{numeric. Shape parameter of the gamma distribution representing the delay between symtom onset and observation.}

\item{parm2_onset_to_report}{numeric. Scale parameter of the gamma distribution representing the delay between symtom onset and observation.}

\item{distribution_type_incubation}{string.}

\item{distribution_type_onset_to_report}{string.}

\item{max_quantile}{numeric. between 0 and 1.}
}
\value{
vector specifying the CDF between each time step of the waiting time distribution.
}
\description{
Build a waiting time distribution from the convolution of two gamma distributions
}
\examples{
#TODO add examples
}
