% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{convolve_delay_inputs}
\alias{convolve_delay_inputs}
\title{Convolve delay inputs.}
\usage{
convolve_delay_inputs(
  delay_incubation,
  delay_onset_to_report,
  n_report_time_steps,
  ...
)
}
\arguments{
\item{delay_incubation}{Incubation delay. Flexible format.}

\item{delay_onset_to_report}{Delay between symptom onset and report. Flexible format.}

\item{n_report_time_steps}{integer. Length of incidence time series.
Use only when providing empirical delay data.}

\item{...}{
  Arguments passed on to \code{\link[=get_matrix_from_empirical_delay_distr]{get_matrix_from_empirical_delay_distr}}
  \describe{
    \item{\code{min_number_cases}}{integer. Minimal number of cases to build empirical distribution from}
    \item{\code{upper_quantile_threshold}}{numeric. Between 0 and 1. TODO add details}
    \item{\code{ref_date}}{Date. Optional. Date of the first data entry in \code{incidence_data}}
    \item{\code{time_step}}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}
  }}
}
\value{
a discretized delay distribution vector or matrix.
}
\description{
This function is flexible in the type of delay inputs it can handle.
Each delay input can be one of:
\itemize{
\item{a list representing a distribution object}
\item{a discretized delay distribution vector}
\item{a discretized delay distribution matrix}
\item{a dataframe containing empirical delay data}
}
}
\details{
see \code{\link{get_matrix_from_empirical_delay_distr}} for details on the format
expected for the empirical delay data.
}
