% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{dating}
\alias{dating}
\title{Dating parameters}
\arguments{
\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}
}
\description{
Dating parameters
}
