% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve_incidence_Richardson_Lucy}
\alias{deconvolve_incidence_Richardson_Lucy}
\title{Title}
\usage{
deconvolve_incidence_Richardson_Lucy(
  incidence_input,
  delay_distribution,
  initial_delta,
  time_units_in_the_past = 30,
  threshold_chi_squared = 1,
  max_iterations = 100,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
