% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{deconvolve_using_empirical_delays}
\alias{deconvolve_using_empirical_delays}
\title{Title}
\usage{
deconvolve_using_empirical_delays(
  incidence_vector,
  empirical_delay_data,
  deconvolution_method = "Richardson-Lucy delay distribution",
  shape_incubation,
  scale_incubation,
  ref_date,
  time_step = "day",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
