% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{delay_empirical}
\alias{delay_empirical}
\title{Empirical delays parameters}
\arguments{
\item{empirical_delays, delays}{dataframe containing the empirical data. See Details.}

\item{n_report_time_steps}{integer. Length of the incidence time series in the accompanying analysis.
This argument is needed to determine the dimensions of the output matrix.}

\item{min_number_cases}{integer.
Minimal number of cases to build the empirical distribution from.
If \code{num_steps_in_a_unit} is \code{NULL}, for any time step T,
the \code{min_number_cases} records prior to T are used.
If less than \code{min_number_cases} delays were recorded before T,
then T is ignored and the \code{min_number_cases} earliest-recorded delays are used.
If \code{num_steps_in_a_unit} is given a value, a similar same procedure is applied,
except that, now at least \code{min_number_cases} must be taken over a round number of
time units. For example, if \code{num_steps_in_a_unit = 7}, and time steps represent consecutive days,
to build the distribution for time step T,
we find the smallest number of weeks starting from T and going in the past,
for which at least \code{min_number_cases} delays were recorded.
We then use all the delays recorded during these weeks.
Weeks are not meant as necessarily being Monday to Sunday,
but simply 7 days in a row, e.g. it can be Thursday-Wednesday.
Again, if less than \code{min_number_cases} delays were recorded before T,
then T is ignored.
We then find the minimum number of weeks, starting from the first recorded delay
that contains at least \code{min_number_cases}.}

\item{min_number_cases_fraction}{numeric. Between 0 and 1.
If \code{min_number_cases} is not provided (kept to \code{NULL}),
the number of most-recent cases used to build
the instant delay distribution is \code{min_number_cases_fraction}
times the total number of reported delays.}

\item{min_min_number_cases}{numeric. Lower bound
for number of cases used to build an instant delay distribution.}

\item{upper_quantile_threshold}{numeric. Between 0 and 1.
Argument for internal use.}

\item{fit}{string. One of "gamma" or "none". Specifies the type of fit that
is applied to the columns of the delay matrix}

\item{date_of_interest}{Date. Date for which the most recent recorded delays are sought.}

\item{num_steps_in_a_unit}{Optional argument.
Number of time steps in a full time unit (e.g. 7 if looking at weeks).
If set, the delays used to build a particular
delay distribution will span over a round number of such time units.
This option is included for comparison with legacy code.}
}
\description{
Empirical delays parameters
}
