% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{distribution}
\alias{distribution}
\title{Distribution}
\arguments{
\item{distribution}{list. probability distribution specified in list format
e.g. list(name = "gamma", shape = 2, scale = 4).
The \code{distribution} list must contain a 'name' element, this element must be  a string and
correspond to one of the types of \code{\link[stats:Distributions]{distributions}} supported in the \link[stats]{Distributions} package.
\code{distribution} must also contain parameters for the specified distribution, in the form '\code{parameter_name=parameter_value}'.}
}
\description{
Distribution
}
