% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_date_column}
\alias{.add_date_column}
\title{Add dates column to dataframe.}
\usage{
.add_date_column(
  estimates,
  ref_date,
  time_step,
  index_col = "idx",
  keep_index_col = FALSE
)
}
\arguments{
\item{estimates}{dataframe. Estimates.}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{index_col}{string. Index column to keep track of which data point
in bootstrapped estimates corresponds to which data point in the original estimates.}

\item{keep_index_col}{boolean. Keep index column in output?}
}
\value{
estimates dataframe with dates column.
}
\description{
Add dates column to dataframe.
}
