% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{.add_noise}
\alias{.add_noise}
\title{Add noise to a series of observations.}
\usage{
.add_noise(observations, noise = list(type = "iid_noise_sd", sd = 1))
}
\arguments{
\item{observations}{Numeric vector. Series of observations through time.}

\item{noise}{List specifying the type of noise and its parameters, if applicable.}
}
\value{
Positive integer vector. Noisy observations.
}
\description{
Add noise to a series of observations.
}
