% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.are_valid_argument_values}
\alias{.are_valid_argument_values}
\title{Utility functions for input validity.}
\usage{
.are_valid_argument_values(user_inputs)
}
\arguments{
\item{user_inputs}{A list of lists with two elements: the first is the value of the parameter to be tested. The second is the expected type of that parameter.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function that checks that the values the user passed when calling a function are valid.
}
