% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.block_bootstrap}
\alias{.block_bootstrap}
\title{Apply block-bootstrapping procedure to module input}
\usage{
.block_bootstrap(
  incidence_input,
  round_incidence = TRUE,
  smoothing_method = "LOESS",
  ...
)
}
\arguments{
\item{incidence_input}{module input object.}

\item{round_incidence}{boolean. Round the bootstrapped incidence?}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}

\item{...}{
  Arguments passed on to \code{\link[=.block_bootstrap_overlap_func]{.block_bootstrap_overlap_func}}, \code{\link[=smooth_incidence]{smooth_incidence}}
  \describe{
    \item{\code{block_size}}{integer. Size of a bootstrapping block.}
    \item{\code{keep_weekdays_aligned}}{boolean. Set to FALSE if not daily incidence, or if no weekly noise pattern.}
  }}
}
\value{
a module output object. bootstrapped incidence.
}
\description{
\code{.block_bootstrap} returns a block-bootstrapped replicate
of the incidence. Incidence should be a vector of non-negative values
}
\details{
This function works by resampling blocks of differences (on the log-scale)
between the original data and a smoothed version of the original data.
}
