% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.block_bootstrap_overlap_func}
\alias{.block_bootstrap_overlap_func}
\title{Helper function for block-bootstrapping}
\usage{
.block_bootstrap_overlap_func(
  incidence_vector,
  block_size = 10,
  keep_weekdays_aligned = TRUE
)
}
\arguments{
\item{incidence_vector}{module input. Original incidence to bootstrap over.}

\item{block_size}{integer. Size of a bootstrapping block.}

\item{keep_weekdays_aligned}{boolean. Set to FALSE if not daily incidence, or if no weekly noise pattern.}
}
\value{
numeric vector. Bootstrapped differences.
}
\description{
Builds a bootstrapped vector of errors.
}
