% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{.build_list_of_gradually_changing_delays}
\alias{.build_list_of_gradually_changing_delays}
\title{TODO doc}
\usage{
.build_list_of_gradually_changing_delays(init_delay, final_delay, n_time_steps)
}
\arguments{
\item{n_time_steps}{}
}
\value{

}
\description{
TODO doc
}
