% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.check_class}
\alias{.check_class}
\title{Check whether the class of an object is as expected}
\usage{
.check_class(object, proper_class)
}
\arguments{
\item{object}{An object whose class needs checking,}

\item{proper_class}{A string describing the desired class of 'object'.}
}
\value{
TRUE if no error is thrown.
}
\description{
Check whether the class of an object is as expected
}
