% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.check_if_positive_number}
\alias{.check_if_positive_number}
\title{Utility functions for input validity.}
\usage{
.check_if_positive_number(number, parameter_name)
}
\arguments{
\item{number}{The value to be tested}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function to check whether an object is a strictly positive number
}
