% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_is_delay_distribution_matrix}
\alias{.check_is_delay_distribution_matrix}
\title{Utility functions for input validity.}
\usage{
.check_is_delay_distribution_matrix(delay_matrix, incidence_data_length)
}
\arguments{
\item{delay_matrix}{A matrix to be tested}

\item{incidence_data_length}{A number representing the length of the given incidence data.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function that checks if a given matrix is a valid delay distribution matrix.
For this, the matrix needs to fulfill the following conditions:
\itemize{
\item is a numeric matrix
\item has no values < 0
\item is a lower triangular matrix
\item no column sums up to more than 1
\item no NA values
\item the size of the matrix is greater than the length of the incidence data
}
}
