% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.check_is_numeric_in_interval}
\alias{.check_is_numeric_in_interval}
\title{Utility functions for input validity.}
\usage{
.check_is_numeric_in_interval(
  user_input,
  parameter_name,
  interval_start,
  interval_end
)
}
\arguments{
\item{user_input}{The variable to be tested}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}

\item{interval_start}{Left-bound of the accepted interval}

\item{interval_end}{Right-bound of the accepted interval}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function to check whether an object is a number that belongs to a given interval
}
