% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.check_is_probability_distr_vector}
\alias{.check_is_probability_distr_vector}
\title{Check that an object represents a probability distribution.}
\usage{
.check_is_probability_distr_vector(
  distribution,
  tolerate_NAs = FALSE,
  tolerance_on_sum = 0.001
)
}
\arguments{
\item{distribution}{Input for which we need to check that it is a proper probability distribution.}

\item{tolerate_NAs}{Can the distribution contain NA values?}

\item{tolerance_on_sum}{Numeric tolerance in checking that vector elements sum to 1.}
}
\value{
TRUE if no error was raised.
}
\description{
To pass the check:
\enumerate{
\item the object must be a numeric vector
\item its elements must sum to 1
\item it must not contain any strictly-negative value.
\item (optionally) it must not contain NAs.
}
}
