% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{.compute_discretized_infectiousness}
\alias{.compute_discretized_infectiousness}
\title{Compute the discretized infectiousness for a particular time step.}
\usage{
.compute_discretized_infectiousness(k, shapeG = 2.73, scaleG = 1.39)
}
\arguments{
\item{k}{Integer. Infex of the time step.}

\item{shapeG}{shape of the gamma distribution of the serial interval -1.}

\item{scaleG}{scale of the gamma distribution of the serial interval -1.}
}
\value{
value of the infectiousness profile at time step k.
}
\description{
The gamma distribution specified is not directly the serial interval but the serial interval
minus one, to allow for a gamma distribution (otherwise infectiousness at day 0 could not be handled),
see Cori et al. 2013 for more details.
}
