% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{.compute_Ot}
\alias{.compute_Ot}
\title{Compute the number of delayed observations of infections at the current time step.}
\usage{
.compute_Ot(infections, delay_distribution, day)
}
\arguments{
\item{infections}{Vector representing infections through time.}

\item{delay_distribution}{Numeric vector. Discretized delay distribution represented as a vector.}

\item{day}{Integer. index of the current time step.}
}
\value{
Integer. Number of observations made on a particular time step.
}
\description{
Compute the number of delayed observations of infections at the current time step.
}
