% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distribution_matrices}
\alias{.convolve_delay_distribution_matrices}
\title{Convolve two delay distribution matrices}
\usage{
.convolve_delay_distribution_matrices(matrix_a, matrix_b)
}
\arguments{
\item{matrix_a}{discretized delay distribution matrix}

\item{matrix_b}{discretized delay distribution matrix}
}
\value{
convolved discretized delay distribution matrix
}
\description{
Note that this convolution operation is not commutative!
The order matters: here the delay implied by \code{matrix_a} happens first,
then the one implied by \code{matrix_b}.
}
