% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distribution_vector_with_matrix}
\alias{.convolve_delay_distribution_vector_with_matrix}
\title{Convolve a delay distribution vector with a delay distribution matrix}
\usage{
.convolve_delay_distribution_vector_with_matrix(
  vector_a,
  matrix_b,
  vector_first = TRUE
)
}
\arguments{
\item{vector_a}{discretized delay distribution vector}

\item{matrix_b}{discretized delay distribution matrix}

\item{vector_first}{Is the delay described by \code{vector_a} happen before
the one from \code{matrix_b} or is it the opposite?}
}
\value{
discretized delay distribution matrix
}
\description{
Convolve a delay distribution vector with a delay distribution matrix
}
