% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distribution_vectors}
\alias{.convolve_delay_distribution_vectors}
\title{Convolve two discretized probability distribution vectors.}
\usage{
.convolve_delay_distribution_vectors(vector_a, vector_b)
}
\arguments{
\item{vector_a, vector_b}{Discretized probability distribution vectors}
}
\value{
discretized probability distribution vector
}
\description{
Convolve two discretized probability distribution vectors.
}
