% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distributions}
\alias{.convolve_delay_distributions}
\title{Convolve two delay vectors or matrices}
\usage{
.convolve_delay_distributions(first_delay, second_delay)
}
\arguments{
\item{first_delay}{discretized delay distribution vector or matrix}

\item{second_delay}{discretized delay distribution vector or matrix}
}
\value{
discretized delay distribution vector (if both input delays are
vectors) or matrix
}
\description{
Convolve two delay vectors or matrices
}
