% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.convove_delay_distribution_matrices}
\alias{.convove_delay_distribution_matrices}
\title{Title}
\usage{
.convove_delay_distribution_matrices(matrix_a, matrix_b)
}
\arguments{
\item{matrix_a}{square numeric matrix}

\item{matrix_b}{square numeric matrix}
}
\value{
square matrix. Convolved matrix of time-varying delays.
}
\description{
Note that this convolution operation is not commutative!
}
