% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.delay_distribution_matrix_rmse_compute}
\alias{.delay_distribution_matrix_rmse_compute}
\title{Utility function that generates delay data, assuming a different delay between event and observation for each individual day.
It then generates the delay matrix and computes the RMSE between the parameters of the gamma distributions passed as arguments and the ones recovered from the delay matrix.
The shapes and scales of the gamma distributions are specified as parameters, and the number of timesteps is assumed to be equal to the length of these vectors.}
\usage{
.delay_distribution_matrix_rmse_compute(
  original_distribution_shapes,
  original_distribution_scales,
  nr_distribution_samples = 500
)
}
\arguments{
\item{original_distribution_shapes}{vector. Specifies the shapes for the gamma distributions.}

\item{original_distribution_scales}{vector. Specifies the scales for the gamma distributions.}

\item{nr_distribution_samples}{integer. How many cases to be sampled for each timestep.}
}
\value{
A list with the computed RMSE. It has two elements: $shape_rmse and $scale_rmse
}
\description{
Utility function that generates delay data, assuming a different delay between event and observation for each individual day.
It then generates the delay matrix and computes the RMSE between the parameters of the gamma distributions passed as arguments and the ones recovered from the delay matrix.
The shapes and scales of the gamma distributions are specified as parameters, and the number of timesteps is assumed to be equal to the length of these vectors.
}
