% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{.draw_It}
\alias{.draw_It}
\title{Draw the number of infections for a particular time step.}
\usage{
.draw_It(Rt, incidence, day, infectiousness_profile)
}
\arguments{
\item{Rt}{Numeric vector. Reproductive number values through time.}

\item{incidence}{Numeric vector. Incidence of infections through time before the current time step.}

\item{day}{Integer. Index of the current time step.}

\item{infectiousness_profile}{Numeric vector. Discretized infectiousness values through time.}
}
\value{
Positive integer. Simulated number of infections for the current time step.
}
\description{
Draw the number of infections for a particular time step.
}
