% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Re.R
\name{.estimate_Re_EpiEstim_piecewise_constant}
\alias{.estimate_Re_EpiEstim_piecewise_constant}
\title{Estimate Re with EpiEstim in a piecewise-constant fashion}
\usage{
.estimate_Re_EpiEstim_piecewise_constant(
  incidence_input,
  minimum_cumul_incidence = 10,
  interval_ends = NULL,
  interval_length = 7,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1,
  output_HPD = FALSE
)
}
\arguments{
\item{incidence_input}{module input object.}

\item{minimum_cumul_incidence}{Numeric scalar. Minimum number of cumulated infections before starting the Re estimation}

\item{mean_serial_interval}{Numeric positive scalar. \code{mean_si} for \code{\link[EpiEstim]{estimate_R}}}

\item{std_serial_interval}{Numeric positive scalar. \code{std_si} for \code{\link[EpiEstim]{estimate_R}}}

\item{mean_Re_prior}{Numeric positive scalar. \code{mean prior} for \code{\link[EpiEstim]{estimate_R}}}
}
\value{
module output object. mean of Re estimates
}
\description{
Estimate Re with EpiEstim in a piecewise-constant fashion
}
