% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-arguments.R
\name{.get_arg_names}
\alias{.get_arg_names}
\title{Get the names of the arguments of a function}
\usage{
.get_arg_names(func)
}
\arguments{
\item{func}{function}
}
\value{
names of the arguments of \code{func}
}
\description{
Get the names of the arguments of a function
}
