% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-empirical_delays.R
\name{.get_delay_matrix_column}
\alias{.get_delay_matrix_column}
\title{Build a specific column of the delay distribution matrix}
\usage{
.get_delay_matrix_column(
  recent_counts_distribution,
  fit = "none",
  col_number,
  N,
  return_fitted_distribution = FALSE
)
}
\arguments{
\item{recent_counts_distribution}{numeric vector of report delays, as used in \code{get_matrix_from_empirical_delay_distr}}

\item{fit}{string. Can be either "none" or "gamma". Specifies the type of fitting applied to the computed column}

\item{col_number}{positive integer. The index the computed column has in the delay matrix}

\item{N}{positive integer. Size of delay matrix.}

\item{return_fitted_distribution}{boolean. If TRUE, the function also returns the gamma distribution that was fitted to the respective column.}
}
\value{
If \code{return_fitted_distribution = FALSE},
returns the \code{col_number}th column of the delay matrix, based on the vector of report delays given.
If \code{return_fitted_distribution = TRUE}, it returns a list with two elements:
\code{column} - delay matrix column as described above,
and \code{distribution} - the delay distribution that was fitted to the column.
}
\description{
Build a specific column of the delay distribution matrix
}
