% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.get_discretized_distribution}
\alias{.get_discretized_distribution}
\title{Title}
\usage{
.get_discretized_distribution(distribution, right_boundary, offset_by_one)
}
\arguments{
\item{offset_by_one}{}
}
\value{

}
\description{
Title
}
