% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.get_distribution_function}
\alias{.get_distribution_function}
\title{Title}
\usage{
.get_distribution_function(distribution, function_prefix)
}
\arguments{
\item{function_prefix}{}
}
\value{

}
\description{
Title
}
