% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_dots_as_list}
\alias{.get_dots_as_list}
\title{Return dots arguments as list.}
\usage{
.get_dots_as_list(...)
}
\arguments{
\item{...}{dots arguments.}
}
\value{
list containing dots arguments or empty list
}
\description{
If there are no dots arguments, then return an empty list.
}
