% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{.get_infectiousness_profile}
\alias{.get_infectiousness_profile}
\title{Compute a discretized infectiousness profile from a serial interval distribution.}
\usage{
.get_infectiousness_profile(
  mean_SI = 4.8,
  sd_SI = 2.3,
  stopping_threshold = 0.999999
)
}
\arguments{
\item{mean_SI}{Numeric positive value. Mean of serial interval of epidemic simulated.}

\item{sd_SI}{Numeric positive value. Standard deviation of serial interval of epidemic simulated.}

\item{stopping_threshold}{Numeric value between 0 and 1.
Threshold on cumulative sum of infectiousness profile returned.
There is little reason to change to something else than the default value.}
}
\value{
Discretized infectiousness profile through time,
represented as a numeric vector with the first value being the infectiousness at time step 0
and each subsequent value being the infectiousness on the time step after the previous value.
}
\description{
The serial interval is assumed to be gamma-distributed here,
following Cori et al. 2013.
}
