% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.get_initial_deconvolution_shift}
\alias{.get_initial_deconvolution_shift}
\title{Get initial shift for deconvolution step}
\usage{
.get_initial_deconvolution_shift(delay_distribution_vector)
}
\arguments{
\item{delay_distribution_vector}{}
}
\value{

}
\description{
Get initial shift for deconvolution step
}
