% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_input_length}
\alias{.get_input_length}
\title{Get length of values vector in a module input object.}
\usage{
.get_input_length(input)
}
\arguments{
\item{input}{module input object.}
}
\value{
integer. length of values vector.
}
\description{
Get length of values vector in a module input object.
}
