% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.get_matrix_from_waiting_time_distributions}
\alias{.get_matrix_from_waiting_time_distributions}
\title{Build delay distribution matrix from list of delay distribution vectors}
\usage{
.get_matrix_from_waiting_time_distributions(waiting_time_distribution_list)
}
\arguments{
\item{waiting_time_distribution_list}{}
}
\value{

}
\description{
Build delay distribution matrix from list of delay distribution vectors
}
