% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_module_input}
\alias{.get_module_input}
\title{Transform input data into a module input object}
\usage{
.get_module_input(data)
}
\arguments{
\item{data}{}
}
\value{
module input object
}
\description{
#TODO specify what format data can take
}
