% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.get_module_output}
\alias{.get_module_output}
\title{Transform a result output of a module into a module output 'object'}
\usage{
.get_module_output(results, input, offset = 0)
}
\arguments{
\item{results}{numeric vector containing output of module operations.}

\item{offset}{integer. Shift resulting from operations performed in module.}
}
\description{
Also takes the module input object \code{input}
given to the module to calculate the offset of the output object.
The new offset is simply (module input offset) + (shift applied during module operations)
The shift applied during module operations is the \code{offset} argument.
}
