% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_offset}
\alias{.get_offset}
\title{Get offset from module object}
\usage{
.get_offset(module_object)
}
\arguments{
\item{module_object}{}
}
\value{
numeric scalar. Offset
}
\description{
This function must be adapted if the module_input_object implementation changes.
}
