% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_shared_args}
\alias{.get_shared_args}
\title{Get the arguments which apply to a function among a given list of arguments.}
\usage{
.get_shared_args(func_list, dots_args)
}
\arguments{
\item{func_list}{list of functions}

\item{dots_args}{list of arguments}
}
\value{
elements of \code{dots_args} which can be passed to \code{func_list}
}
\description{
This function is used to find which arguments should be passed
to a list of functions among the dot arguments passed to a higher-level function.
}
