% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.get_time_steps_quantile}
\alias{.get_time_steps_quantile}
\title{Get initial shift for deconvolution step}
\usage{
.get_time_steps_quantile(delay_distribution_vector, quantile = 0.5)
}
\arguments{
\item{delay_distribution_vector}{discretized probability distribution vector}

\item{quantile}{Value between 0 and 1. Quantile of the distribution.}
}
\value{
an integer value corresponding to the rounded value
corresponding to the specified quantile of the input delay distribution.
The returned calue is always non-negative.
}
\description{
Return the time step corresponding to a specified quantile
for a particular distribution (specified as a vector).
}
