% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_values}
\alias{.get_values}
\title{Get values from module object}
\usage{
.get_values(module_object)
}
\arguments{
\item{module_input_object}{}
}
\value{
vector containing values
}
\description{
This function must be adapted if the module_input_object implementation changes.
}
