% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.get_vector_constant_waiting_time_distr}
\alias{.get_vector_constant_waiting_time_distr}
\title{Build a waiting time distribution from the convolution of two gamma distributions}
\usage{
.get_vector_constant_waiting_time_distr(
  shape_incubation,
  scale_incubation,
  shape_onset_to_report,
  scale_onset_to_report,
  length_out = 200,
  n_random_samples = 1e+06
)
}
\arguments{
\item{shape_incubation}{numeric. Shape parameter of the gamma distribution representing the delay between infection and symptom onset.}

\item{scale_incubation}{numeric. Scale parameter of the gamma distribution representing the delay between infection and symptom onset.}

\item{shape_onset_to_report}{numeric. Shape parameter of the gamma distribution representing the delay between symtom onset and observation.}

\item{scale_onset_to_report}{numeric. Scale parameter of the gamma distribution representing the delay between symtom onset and observation.}

\item{length_out}{integer. Length of the output vector}

\item{n_random_samples}{integer. number of samples drawn from each distribution}
}
\value{
vector specifying the CDF between each time step of the waiting time distribution.
}
\description{
Build a waiting time distribution from the convolution of two gamma distributions
}
