% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.is_numeric_vector}
\alias{.is_numeric_vector}
\title{Check if object is numeric vector.}
\usage{
.is_numeric_vector(object)
}
\arguments{
\item{object}{Any object.}
}
\value{
TRUE if numeric vector, FALSE otherwise
}
\description{
Check if object is numeric vector.
}
