% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.is_positive_numeric_vector}
\alias{.is_positive_numeric_vector}
\title{Utility functions for input validity.}
\usage{
.is_positive_numeric_vector(vector)
}
\arguments{
\item{vector}{vector to be tested}
}
\value{
boolean. TRUE if vector is a positive numeric vector. FALSE otherwise
}
\description{
Utility function to determine whether an object is a numeric vector with all positive (or zero) values.
}
