% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.is_single_delay}
\alias{.is_single_delay}
\title{Is input a single delay object?}
\usage{
.is_single_delay(delay_list)
}
\arguments{
\item{delay_list}{list or single delay object}
}
\value{
TRUE if single delay, FALSE otherwise
}
\description{
Is input a single delay object?
}
