% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.is_valid_delay_object}
\alias{.is_valid_delay_object}
\title{Utility functions for input validity.}
\usage{
.is_valid_delay_object(delay_object, parameter_name, incidence_data_length)
}
\arguments{
\item{delay_object}{user inputted object to be tested}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}

\item{incidence_data_length}{A number representing the length of the given incidence data.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function that checks whether a user input is a valid delay object. This means it can be one of the following:
\itemize{
\item a probability distribution vector: a numeric vector with no \code{NA} or negative values, whose entries sum up to 1
\item an empirical delay data: a data frame with two columns: \code{event_date} and \code{report_delay}. The columns cannot contain \code{NA} values. \code{report_delay} only contains non-negative values
\item a delay distribution matrix (as described in \code{\link{.check_is_delay_distribution_matrix}})
\item a distribution object (e.g. list(name = 'gamma', scale = X, shape = Y))
}
}
