% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_valid_module_input}
\alias{.is_valid_module_input}
\title{Utility functions for input validity.}
\usage{
.is_valid_module_input(module_input_object, parameter_name)
}
\arguments{
\item{module_input_object}{the vector/list the user passed as a parameter, to be tested}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function that checks if a user input is one of:
\itemize{
\item a numeric vector with values > 0
\item a list with two elements: \code{values} (a numeric vector with values > 0) and \code{index_offset} (an integer)
}
}
