% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.is_value_in_accepted_values_vector}
\alias{.is_value_in_accepted_values_vector}
\title{Utility functions for input validity.}
\usage{
.is_value_in_accepted_values_vector(string_user_input, parameter_name)
}
\arguments{
\item{string_user_input}{A string containing the value that the user passed for the tested string type parameter.}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function that checks if a specific user given parameter value is among the accepted ones, in which case it returns TRUE
Throws an error otherwise.
}
