% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_value_valid_time_step}
\alias{.is_value_valid_time_step}
\title{Utility functions for input validity.}
\usage{
.is_value_valid_time_step(string_user_input, parameter_name)
}
\arguments{
\item{string_user_input}{A string containing the value that the user passed for the tested string type parameter.}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function that checks if a specific user given parameter value an accepted time_step, in which case it returns TRUE
An accepted time_step is considered to be:
<<A character string, containing one of "day", "week", "month", "quarter" or "year".
This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".>>
(from \url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/seq.Date})
}
