% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.left_augment_delay_distribution}
\alias{.left_augment_delay_distribution}
\title{Title}
\usage{
.left_augment_delay_distribution(delay_distribution_matrix, n_col_augment)
}
\arguments{
\item{n_col_augment}{an integer.}
}
\value{

}
\description{
Title
}
