% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.make_ecdf_from_empirical_data_and_gamma}
\alias{.make_ecdf_from_empirical_data_and_gamma}
\title{Build an empirical Cumulative Distribution Function
from the convolution of a gamma distribution and an empirical distribution}
\usage{
.make_ecdf_from_empirical_data_and_gamma(
  gamma_draws,
  empirical_distr,
  multiplier_init = 100
)
}
\arguments{
\item{shape}{numeric vector}

\item{scale}{numeric vector}

\item{number_of_samples}{integer. Number of samples.}
}
\value{
ecdf object
}
\description{
Utility function that sums samples from an empirical distribution and a gamma distribution.
}
