% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_empty_module_output}
\alias{.make_empty_module_output}
\title{Make empty module output object}
\usage{
.make_empty_module_output()
}
\value{
empty module output object
}
\description{
Make empty module output object
}
