% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_tibble_from_output}
\alias{.make_tibble_from_output}
\title{Convert module output object into tibble}
\usage{
.make_tibble_from_output(output, output_name)
}
\arguments{
\item{output}{module output object}

\item{output_name}{string. Name to be given to the values column}
}
\value{
tibble
}
\description{
Convert module output object into tibble
}
