% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.prettify_result}
\alias{.prettify_result}
\title{Prettify results of pipe functions by removing leading and tailing NAs}
\usage{
.prettify_result(data, index_col = "idx", date_col = "date")
}
\arguments{
\item{data}{Module object or dataframe.
If dataframe, must contain a column named \code{index_col}
or \code{date_col} (or both),}

\item{index_col}{string. Name of the index column.}

\item{date_col}{string. Name of the date column.}
}
\value{
The input dataframe without leading NA rows.
}
\description{
Prettify results of pipe functions by removing leading and tailing NAs
}
