% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testing.R
\name{.print_vector}
\alias{.print_vector}
\title{Utility function to print vectors in a copy-pastable format}
\usage{
.print_vector(a, digits = 2)
}
\arguments{
\item{a}{numeric vector}

\item{digits}{integer. Number of digits to round.}
}
\value{
Print vector as string
}
\description{
Utility function to print vectors in a copy-pastable format
}
