% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{.smooth_LOESS}
\alias{.smooth_LOESS}
\title{LOESS smoothing function}
\usage{
.smooth_LOESS(incidence_input, data_points_incl = 21, degree = 1)
}
\arguments{
\item{incidence_input}{module input object.}

\item{data_points_incl}{integer. Size of the window used in the LOESS algorithm.}

\item{degree}{integer. LOESS degree.}
}
\value{
module output object. Smoothed incidence.
}
\description{
Prefer the use of the wrapper function \code{smooth_incidence(..., smoothing_method = "LOESS")}
instead of \code{.smooth_LOESS}.
}
\details{
This function implements the LOESS method for smoothing noisy data.
It is essentially a wrapper around \code{\link[stats]{loess}}.
See the help section for \code{\link[stats]{loess}} for details on LOESS.

#TODO add details on how data_points_incl relates to span.
}
