% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{.summarise_bagged_mean}
\alias{.summarise_bagged_mean}
\title{Title}
\usage{
.summarise_bagged_mean(
  original_estimates,
  bootstrapped_estimates,
  Re_estimate_col = Re_estimate_col,
  bootstrap_id_col = bootstrap_id_col,
  time_step = "day"
)
}
\arguments{
\item{bootstrap_id_col}{}
}
\value{

}
\description{
Title
}
