% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{EpiEstim_wrapper}
\alias{EpiEstim_wrapper}
\title{EpiEstim wrappers arguments}
\arguments{
\item{minimum_cumul_incidence}{Numeric value.
Minimum number of cumulated infections before starting the Re estimation.
Default is \code{12} as recommended in Cori et al., 2013.}

\item{mean_serial_interval}{Numeric positive value. \code{mean_si} for \code{\link[EpiEstim]{estimate_R()}}}

\item{std_serial_interval}{Numeric positive value. \code{std_si} for \code{\link[EpiEstim]{estimate_R()}}}

\item{mean_Re_prior}{Numeric positive value. \code{mean prior} for \code{\link[EpiEstim]{estimate_R()}}}
}
\value{
If \code{output_HPD = FALSE},
value is a module object (a list of the same kind as \code{incidence_input}).
The \code{values} element of the list then contains the Re estimates.
If \code{output_HPD = TRUE}, a list of three module objects is returned.
\itemize{
\item \code{Re_estimate} contains the Re estimates.
\item \code{Re_highHPD} and \code{Re_lowHPD} contain
the higher and lower boundaries of the HPD interval,
as computed by \code{\link[EpiEstim]{estimate_R()}}
}
}
\description{
EpiEstim wrappers arguments
}
