% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EST_incidence_data}
\alias{EST_incidence_data}
\title{Incidence data for COVID-19 in Estonia}
\format{
A data frame with 460 rows and 2 variables:
\describe{
\item{date}{date of case reporting in YYYY-mm-dd format}
\item{case_incidence}{number of cases reported on this date}
}
}
\source{
\url{https://opendata.digilugu.ee/opendata_covid19_tests_total.csv}
}
\usage{
EST_incidence_data
}
\description{
A dataset containing aggregated incidence data for Estonia from February 2020 to May 2021.
The linelist data is published by the Estonian public health authorities.
}
\keyword{datasets}
