% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{estimate_from_combined_observations}
\alias{estimate_from_combined_observations}
\title{Title}
\usage{
estimate_from_combined_observations(
  partially_delayed_incidence,
  fully_delayed_incidence,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  delay_until_partial,
  delay_from_partial_to_full,
  partial_observation_requires_full_observation = TRUE,
  ref_date = NULL,
  time_step = "day",
  output_Re_only = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
