% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Re.R
\name{estimate_Re_EpiEstim_sliding_window}
\alias{estimate_Re_EpiEstim_sliding_window}
\title{Title}
\usage{
estimate_Re_EpiEstim_sliding_window(
  incidence_input,
  minimul_cumul_incidence = 0,
  estimation_window = 3,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1
)
}
\arguments{
\item{minimul_cumul_incidence}{Numeric scalar. Minimum number of cumulated infections before starting the Re estimation}

\item{mean_Re_prior}{}
}
\value{

}
\description{
Title
}
