% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testing.R
\name{expect_delay_matrix_sums_lte_1}
\alias{expect_delay_matrix_sums_lte_1}
\title{Test whether a delay matrix columns sum to less than one.}
\usage{
expect_delay_matrix_sums_lte_1(matrix, full_cols = 0, tolerance = 0.001)
}
\arguments{
\item{matrix}{input matrix}

\item{full_cols}{number of full columns.
Full columns must sum to 1.}

\item{tolerance}{tolerance on the result}
}
\value{
TRUE if test passed
}
\description{
Test whether a delay matrix columns sum to less than one.
}
