% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_delay_data}
\alias{generate_delay_data}
\title{Generate delay data.}
\usage{
generate_delay_data(
  origin_date = as.Date("2020-02-01"),
  n_time_steps = 100,
  time_step = "day",
  delay_ratio_start_to_end = 2,
  distribution_initial_delay = list(name = "gamma", shape = 6, scale = 5),
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\value{
data.frame. Simulated delay data.
}
\description{
This utility can be used to build toy examples to test functions dealing with empirical delay data.
}
\examples{
#TODO add examples
}
