% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_block_bootstrapped_estimate}
\alias{get_block_bootstrapped_estimate}
\title{Title}
\usage{
get_block_bootstrapped_estimate(
  incidence_vector,
  N_bootstrap_replicates = 100,
  smoothing_method = "LOESS",
  shape_incubation,
  scale_incubation,
  shape_onset_to_report,
  scale_onset_to_report,
  length_out = 200,
  n_random_samples = 1e+06,
  minimul_cumul_incidence = 0,
  estimation_window = 3,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1
)
}
\arguments{
\item{mean_Re_prior}{}
}
\value{

}
\description{
Title
}
