% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_bootstrapped_estimates_from_combined_observations}
\alias{get_bootstrapped_estimates_from_combined_observations}
\title{Title}
\usage{
get_bootstrapped_estimates_from_combined_observations(
  partially_delayed_incidence,
  fully_delayed_incidence,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  bootstrapping_method = "non-parametric block boostrap",
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  N_bootstrap_replicates = 100,
  delay_until_partial,
  delay_from_partial_to_full,
  partial_observation_requires_full_observation = TRUE,
  ref_date = NULL,
  time_step = "day",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
