% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{get_matrix_empirical_waiting_time_distr}
\alias{get_matrix_empirical_waiting_time_distr}
\title{Build matrix of delay distributions through time from empirical delay data.}
\usage{
get_matrix_empirical_waiting_time_distr(
  empirical_delays,
  start_date,
  N,
  time_step = "day",
  min_number_cases = 300,
  upper_quantile_threshold = 0.99
)
}
\arguments{
\item{empirical_delays}{tibble. format to be specified}

\item{start_date}{Date. First date of incidence data}

\item{N}{integer. Length of incidence time series}

\item{time_step}{string. "day", "X days", "week", "month"... (see \link{\code{seq.Date}} for details)}

\item{min_number_cases}{integer. Minimal number of cases to build empirical distribution from}

\item{upper_quantile_threshold}{numeric. Between 0 and 1. TODO add details}
}
\value{

}
\description{
This matrix is required for the application of the Richardson-Lucy algorithm.
}
\examples{
#TODO add example
}
