% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{get_matrix_from_constant_waiting_time_distr}
\alias{get_matrix_from_constant_waiting_time_distr}
\title{Make square delay distribution matrix from vector of delay distributions}
\usage{
get_matrix_from_constant_waiting_time_distr(waiting_time_distr, N)
}
\arguments{
\item{waiting_time_distr}{numeric vector}

\item{N}{integer. Dimension of output matrix}
}
\value{
square numeric matrix
}
\description{
Make square delay distribution matrix from vector of delay distributions
}
