% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{get_matrix_from_empirical_delay_distr}
\alias{get_matrix_from_empirical_delay_distr}
\title{Build matrix of delay distributions through time from empirical delay data.}
\usage{
get_matrix_from_empirical_delay_distr(
  empirical_delays,
  n_report_time_steps,
  start_date = NULL,
  time_step = "day",
  min_number_cases = NULL,
  upper_quantile_threshold = 0.99,
  min_number_cases_fraction = 0.05,
  min_min_number_cases = 10
)
}
\arguments{
\item{empirical_delays}{tibble. format to be specified}

\item{n_report_time_steps}{integer. Length of incidence time series}

\item{start_date}{Date. First date of incidence data}

\item{time_step}{string. "day", "X days", "week", "month"... (see \code{\link[base]{seq.Date}} for details)}

\item{min_number_cases}{integer. Minimal number of cases to build empirical distribution from}

\item{upper_quantile_threshold}{numeric. Between 0 and 1. TODO add details}
}
\value{

}
\description{
This matrix is required for the application of the Richardson-Lucy algorithm.
}
\examples{
#TODO add example
}
