% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HK_incidence_data}
\alias{HK_incidence_data}
\title{Incidence data for COVID-19 in Hong Kong}
\format{
A data frame with 196 rows and 4 variables:
\describe{
\item{date}{date of case reporting in YYYY-mm-dd format}
\item{case_incidence}{total number of case confirmations
reported on this date}
\item{onset_incidence}{number of events of onset of symptoms
occurring on this date}
\item{report_incidence}{number of case confirmations reported on this date
with no known date of onset of symptoms (or asymptomatoc cases)}
}
}
\source{
\url{https://www.chp.gov.hk}
}
\usage{
HK_incidence_data
}
\description{
A dataset containing aggregated incidence data for Hong Kong from January 2020 to July 2021.
This data was  put together by aggregating the publicly-available linelist data for SARS-CoV-2 in Hong Kong.
The linelist data is published by the Centre for Health Protection in Hong Kong.
}
\keyword{datasets}
