% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{inner_module}
\alias{inner_module}
\title{Inner module option characteristics}
\arguments{
\item{incidence_input, input, output}{Module input object.
List with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}}
}
\description{
Inner module option characteristics
}
