% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{loss_function}
\alias{loss_function}
\title{Loss Function}
\usage{
loss_function(beta, X, y)
}
\arguments{
\item{beta}{A \code{vector} of coefficients associated with the features in xi.}

\item{X}{A \code{string} containing the input data for the ith observation}

\item{y}{A \code{numeric} (integer) used to denote the number of simulations.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{I}{Estimated value of the integral}
     \item{var}{Estimated variance of the estimator}
}
}
\description{
Calculates the negative log likelihood p_i
}
\examples{
loss_function([1,2], [1,2], 1)
}
\author{
Stephane Guerrier
}
