% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{make_ecdf_from_two_gammas}
\alias{make_ecdf_from_two_gammas}
\title{Build an empirical Cumulative Distribution Function
from the convolution of two independent Gamma distributions}
\usage{
make_ecdf_from_two_gammas(shape, scale, number_of_samples = 1e+06)
}
\arguments{
\item{shape}{numeric vector}

\item{scale}{numeric vector}

\item{number_of_samples}{integer. Number of samples.}
}
\value{
Vectorize ecdf object
}
\description{
Utility function that draws a large number of samples (1E6 by default) from 2 gamma distributions,
sums a sample from each, and buids an empirical cdf from the distribution of samples obtained.
}
