% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{module_objects}
\alias{module_objects}
\title{Module object utilities}
\arguments{
\item{index_col}{string. Index column to keep track of which data point
in bootstrapped estimates corresponds to which data point in the original estimates.}
}
\description{
Module object utilities
}
