% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{module_structure}
\alias{module_structure}
\title{Module structure characteristics}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero? TODO to be described better.}
}
\description{
Module structure characteristics
}
