% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{pipe_params}
\alias{pipe_params}
\title{Pipe parameters}
\arguments{
\item{output_Re_only}{boolean. Should the output only contain Re estimates?
(as opposed to containing results for each intermediate step)}

\item{output_infection_incidence_only}{boolean. TODO add details}
}
\description{
Pipe parameters
}
