% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{simulate}
\alias{simulate}
\title{Simulation parameters}
\arguments{
\item{Rt}{Numeric vector. Reproductive number values through time.}

\item{mean_SI}{Numeric positive value. Mean of serial interval of epidemic simulated.}

\item{sd_SI}{Numeric positive value. Standard deviation of serial interval of epidemic simulated.}

\item{infections}{Positive integer vector. Course of infections through time.}

\item{noise}{List specifying the type of noise and its parameters, if applicable.}
}
\description{
Simulation parameters
}
