% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_combined_observations}
\alias{simulate_combined_observations}
\title{Title}
\usage{
simulate_combined_observations(
  infections,
  delay_until_partial,
  delay_until_final_report,
  prob_partial_observation
)
}
\arguments{
\item{prob_partial_observation}{}
}
\value{

}
\description{
Title
}
