% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_delayed_observations}
\alias{simulate_delayed_observations}
\title{Simulate a series of delayed observations from a series of infections.}
\usage{
simulate_delayed_observations(
  infections,
  delay,
  noise = list(type = "noiseless")
)
}
\arguments{
\item{infections}{Positive integer vector. Course of infections through time.}

\item{delay}{Single delay or list of delays.
Each delay can be one of:
\itemize{
\item{a list representing a distribution object}
\item{a discretized delay distribution vector}
\item{a discretized delay distribution matrix}
\item{a dataframe containing empirical delay data}
}}

\item{noise}{List specifying the type of noise and its parameters, if applicable.}
}
\value{
Integer vector. Simulated delayed observations.
}
\description{
Simulate a series of delayed observations from a series of infections.
}
