% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{smooth_deconvolve_estimate}
\alias{smooth_deconvolve_estimate}
\title{Estimate Re from incidence data}
\usage{
smooth_deconvolve_estimate(
  incidence_data,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  delay_incubation,
  delay_onset_to_report = c(1),
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  ref_date = NULL,
  time_step = "day",
  output_Re_only = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric. TODO define acceptable format}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Options are: "LOESS", implemented in \code{\link{smooth_LOESS}}.}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution).
Options are "Richardson-Lucy delay distribution".}

\item{estimation_method}{string. Method used to estimate reproductive number
values through time from the reconstructed infection timings.
Options are "EpiEstim sliding window".}

\item{mean_serial_interval}{numeric. see \code{\link{estimate_Re}}}

\item{std_serial_interval}{numeric. see \code{\link{estimate_Re}}}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints. "day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{output_Re_only}{boolean. Should the output only contain Re estimates? (as opposed to containing results for each intermediate step)}

\item{verbose}{boolean. Print verbose output?

#TODO add details on output formatting}

\item{estimation_window}{integer. Only use if \code{estimation_method = "EpiEstim sliding window"}.}
}
\value{
effective reproductive number estimates through time.
}
\description{
This pipe function combines a smoothing step using (to remove noise from the original observations),
a deconvolution step (to retrieve infection events from the observed delays),
and an Re estimation step (wrapping around EpiEstim).
}
\details{
The smoothing step uses the LOESS method by default.
The deconvolution step uses the Richardson-Lucy algorithm  by default.
The estimation uses the Cori method with a sliding window  by default.

#TODO clarify input
}
