% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_incidence}
\alias{smooth_incidence}
\title{Smooth Noisy Incidence Data}
\usage{
smooth_incidence(incidence_data, smoothing_method = "LOESS", ...)
}
\arguments{
\item{smoothing_method}{string. Options are "LOESS".}
}
\value{
module output. Smoothed incidence.
}
\description{
Currently only LOESS smoothing (smoothing_method = "LOESS")
is implemented.
}
\details{
#TODO specify input format
}
