% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_LOESS}
\alias{smooth_LOESS}
\title{LOESS smoothing function}
\usage{
smooth_LOESS(incidence_input, days_incl = 21, degree = 1)
}
\arguments{
\item{days_incl}{integer}

\item{degree}{integer}
}
\value{
module output
}
\description{
LOESS smoothing function
}
