% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap}
\alias{block_bootstrap}
\title{Apply block-bootstrapping procedure to module input}
\usage{
block_bootstrap(
  incidence_input,
  block_size = 10,
  data_points_incl = 21,
  degree = 1,
  round_incidence = TRUE
)
}
\arguments{
\item{incidence_input}{module input. Original incidence to bootstrap over.}

\item{block_size}{integer. Size of a bootstrapping block.}

\item{data_points_incl}{integer. Size of the window used in the LOESS algorithm.}

\item{degree}{integer. LOESS degree.}

\item{round_incidence}{boolean. Round the bootstrapped incidence?}
}
\value{
a module output object
}
\description{
\code{.block_bootstrap} returns a block-bootstrapped replicate
of the incidence. Incidence should be a vector of non-negative values
}
\details{
This function works by resampling blocks of differences (on the log-scale)
between the original data and a smoothed version of the original data.
}
