% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{build_delay_distribution}
\alias{build_delay_distribution}
\title{Build a delay distribution vector}
\usage{
build_delay_distribution(
  distribution,
  max_quantile = 0.999,
  offset_by_one = FALSE
)
}
\arguments{
\item{max_quantile}{numeric value between 0 and 1. TODO write what max_quantile does}

\item{offset_by_one}{boolean. Gamma distribution comes from fit on (raw_data + 1) to accommodate zeroes in the raw_data.}
}
\value{
numeric vector.
}
\description{
Build a delay distribution vector
}
\examples{
#TODO add example
}
