% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{combine_incubation_with_reporting_delay}
\alias{combine_incubation_with_reporting_delay}
\title{Build a waiting time distribution from the convolution of two gamma distributions}
\usage{
combine_incubation_with_reporting_delay(
  distribution_incubation,
  distribution_onset_to_report,
  max_quantile = 0.9999,
  incubation_offset_by_one = FALSE,
  onset_to_report_offset_by_one = FALSE
)
}
\arguments{
\item{distribution_incubation}{list. delay between infection and symptom onset}

\item{distribution_onset_to_report}{list. delay between symtom onset and observation}

\item{max_quantile}{numeric. between 0 and 1.}
}
\value{
vector specifying the CDF between each time step of the waiting time distribution.
}
\description{
Build a waiting time distribution from the convolution of two gamma distributions
}
\examples{
#TODO add examples
}
