% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{convolve_delay_inputs}
\alias{convolve_delay_inputs}
\title{Title}
\usage{
convolve_delay_inputs(
  delay_incubation,
  delay_onset_to_report,
  n_report_time_steps = 0,
  start_date = NULL,
  time_step = "day",
  min_number_cases = NULL,
  upper_quantile_threshold = 0.99
)
}
\arguments{
\item{n_report_time_steps}{Integer value required if}

\item{upper_quantile_threshold}{}
}
\value{

}
\description{
Title
}
\examples{
#TODO add example
}
