% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{convolve_delay_inputs}
\alias{convolve_delay_inputs}
\title{Title}
\usage{
convolve_delay_inputs(
  delay_incubation,
  delay_onset_to_report,
  n_report_time_steps = 0,
  start_date = NULL,
  time_step = "day",
  min_number_cases = NULL,
  upper_quantile_threshold = 0.99
)
}
\arguments{
\item{delay_incubation}{Delay}

\item{n_report_time_steps}{Integer value required if}

\item{start_date}{Optional. Date of the first incidence data point.}

\item{time_step}{Size of time step used in delays. String in the format:
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for
details)}

\item{min_number_cases}{Use only if an input is empirical delay data. Number
of cases over which to build empirical delay distribution}

\item{upper_quantile_threshold}{Value between 0 and 1. Quantile}
}
\value{

}
\description{
Title
}
