% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_for_partially_observed_data}
\alias{correct_for_partially_observed_data}
\title{Correct incidence data for yet-to-be-observed fraction of events}
\usage{
correct_for_partially_observed_data(
  incidence_data,
  delay_distribution_final_report,
  cutoff_observation_probability = 0.1,
  ref_date = NULL,
  time_step = "day",
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{delay_distribution_final_report}{TODO refactor to estimateR
Distribution of the delay between the events collected in the incidence data
and the a posteriori observations of these events.}

\item{cutoff_observation_probability}{value between 0 and 1.
Only datapoints for timesteps that have a probability to be observed higher
than \code{cutoff_observation_probability} are kept.
The few datapoints with a lower probability to be observed are trimmed off
the tail of the timeseries.}
}
\value{
module output object
}
\description{
Use this function to correct the tail of an incidence timeseries
if incidence was collected following a subsequent observation event.
For instance, if the incidence represents people starting to show symptoms of a disease
(dates of onset of symptoms), the data would typically have been collected among
individuals whose case was confirmed via a test.
If so, among all events of onset of symptoms, only those who had time to be
confirmed by a test were reported.
Thus, close to the present, there is an underreporting of onset of symptoms events.
In order to account for this effect, this function divides each incidence value
by the probability of an event happening at a particular timestep to have been observed.
Typically, this correction only affects the few most recent datapoints.
}
