% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-partial_observations.R
\name{correct_for_partially_observed_data}
\alias{correct_for_partially_observed_data}
\title{Correct incidence data for yet-to-be-observed fraction of events}
\usage{
correct_for_partially_observed_data(
  incidence_data,
  delay_until_final_report,
  cutoff_observation_probability = 0.33,
  ref_date = NULL,
  time_step = "day",
  ...
)
}
\arguments{
\item{incidence_data}{An object containing incidence data through time.
It can either be:
\itemize{
\item A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}
\item A numeric vector. The vector corresponds to the \code{values} element
descrived above, and \code{index_offset} is implicitely zero.
This means that the first value in \code{incidence_data}
is associated with the reference time step (no shift towards the future or past).
}}

\item{delay_until_final_report}{Single delay or list of delays.
Each delay can be one of:
\itemize{
\item{a list representing a distribution object}
\item{a discretized delay distribution vector}
\item{a discretized delay distribution matrix}
\item{a dataframe containing empirical delay data}
}}

\item{cutoff_observation_probability}{value between 0 and 1.
Only datapoints for timesteps that have a probability of observing a event
higher than \code{cutoff_observation_probability} are kept.
The few datapoints with a lower probability to be observed are trimmed off
the tail of the timeseries.}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{...}{
  Arguments passed on to \code{\link[=get_matrix_from_empirical_delay_distr]{get_matrix_from_empirical_delay_distr}}
  \describe{
    \item{\code{min_number_cases}}{integer.
Minimal number of cases to build the empirical distribution from.
If \code{num_steps_in_a_unit} is \code{NULL}, for any time step T,
the \code{min_number_cases} records prior to T are used.
If less than \code{min_number_cases} delays were recorded before T,
then T is ignored and the \code{min_number_cases} earliest-recorded delays are used.
If \code{num_steps_in_a_unit} is given a value, a similar same procedure is applied,
except that, now at least \code{min_number_cases} must be taken over a round number of
time units. For example, if \code{num_steps_in_a_unit = 7}, and time steps represent consecutive days,
to build the distribution for time step T,
we find the smallest number of weeks starting from T and going in the past,
for which at least \code{min_number_cases} delays were recorded.
We then use all the delays recorded during these weeks.
Weeks are not meant as necessarily being Monday to Sunday,
but simply 7 days in a row, e.g. it can be Thursday-Wednesday.
Again, if less than \code{min_number_cases} delays were recorded before T,
then T is ignored.
We then find the minimum number of weeks, starting from the first recorded delay
that contains at least \code{min_number_cases}.}
    \item{\code{upper_quantile_threshold}}{numeric. Between 0 and 1.
Argument for internal use.}
    \item{\code{min_number_cases_fraction}}{numeric. Between 0 and 1.
If \code{min_number_cases} is not provided (kept to \code{NULL}),
the number of most-recent cases used to build
the instant delay distribution is \code{min_number_cases_fraction}
times the total number of reported delays.}
    \item{\code{min_min_number_cases}}{numeric. Lower bound
for number of cases used to build an instant delay distribution.}
    \item{\code{fit}}{string. One of "gamma" or "none". Specifies the type of fit that
is applied to the columns of the delay matrix}
    \item{\code{num_steps_in_a_unit}}{Optional argument.
Number of time steps in a full time unit (e.g. 7 if looking at weeks).
If set, the delays used to build a particular
delay distribution will span over a round number of such time units.
This option is included for comparison with legacy code.}
    \item{\code{return_fitted_distribution}}{boolean. If TRUE, the function also returns the gamma distribution that was fitted to the respective column.}
  }}
}
\value{
A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the result of the computations on the input data.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the result is shifted compared to an \code{index_offset} of \code{0}.
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
Note that the \code{index_offset} of the output of the function call
accounts for the (optional) \code{index_offset} of the input.
}
If \code{index_offset} is \code{0} and \code{simplify_output = TRUE},
the \code{index_offset} is dropped and the \code{values}
element is returned as a numeric vector.
}
\description{
Use this function to correct the tail of an incidence time series
if incidence was collected following a subsequent observation event.
For instance, if the incidence represents people starting to show symptoms of a disease
(dates of onset of symptoms), the data would typically have been collected among
individuals whose case was confirmed via a test.
If so, among all events of onset of symptoms, only those who had time to be
confirmed by a test were reported.
Thus, close to the present, there is an under-reporting of onset of symptoms events.
In order to account for this effect, this function divides each incidence value
by the probability of an event happening at a particular time step to have been observed.
Typically, this correction only affects the few most recent data points.
A trimming is done at the tail of the time series to avoid correcting for time steps
for which the observation probability is too low, which could result in too uncertain corrected values.
This trimming is tuned via the \code{cutoff_observation_probability} argument.
}
\details{
The \code{ref_date} argument is only needed if the \code{delay_until_final_report}
is passed as a dataframe of individual delay observations (a.k.a empirical delay data).
In that case, \code{ref_date} must correspond to the date of the first time step in \code{incidence_data}.
}
