% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-partial_observations.R
\name{correct_for_partially_observed_data}
\alias{correct_for_partially_observed_data}
\title{Correct incidence data for yet-to-be-observed fraction of events}
\usage{
correct_for_partially_observed_data(
  incidence_data,
  delay_until_final_report,
  cutoff_observation_probability = 0.1,
  ref_date = NULL,
  time_step = "day",
  ...
)
}
\arguments{
\item{incidence_data}{An object containing incidence data through time.
It can either be:
\itemize{
\item A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}
\item A numeric vector. The vector corresponds to the \code{values} element
descrived above, and \code{index_offset} is implicitely zero.
This means that the first value in \code{incidence_data}
is associated with the reference time step (no shift towards the future or past).
}}

\item{cutoff_observation_probability}{value between 0 and 1.
Only datapoints for timesteps that have a probability of observing a event
higher than \code{cutoff_observation_probability} are kept.
The few datapoints with a lower probability to be observed are trimmed off
the tail of the timeseries.}
}
\description{
Use this function to correct the tail of an incidence time series
if incidence was collected following a subsequent observation event.
For instance, if the incidence represents people starting to show symptoms of a disease
(dates of onset of symptoms), the data would typically have been collected among
individuals whose case was confirmed via a test.
If so, among all events of onset of symptoms, only those who had time to be
confirmed by a test were reported.
Thus, close to the present, there is an under-reporting of onset of symptoms events.
In order to account for this effect, this function divides each incidence value
by the probability of an event happening at a particular time step to have been observed.
Typically, this correction only affects the few most recent data points.
A trimming is done at the tail of the time series to avoid correcting for time steps
for which the observation probability is too low, which could result in too uncertain corrected values.
This trimming is tuned via the \code{cutoff_observation_probability} argument.
}
\details{
The \code{ref_date} argument is only needed if the \code{delay_until_final_report}
is passed as a dataframe of individual delay observations (a.k.a empirical delay data).
In that case, \code{ref_date} must correspond to the date of the first time step in \code{incidence_data}.
}
