% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve_incidence}
\alias{deconvolve_incidence}
\title{Infer infection events dates from delayed observations}
\usage{
deconvolve_incidence(
  incidence_data,
  deconvolution_method = "Richardson-Lucy delay distribution",
  delay_incubation,
  delay_onset_to_report = c(1),
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution step).
Available options are:
\itemize{
\item{'Richardson-Lucy delay distribution',
implemented in \code{\link{.deconvolve_incidence_Richardson_Lucy}}}
}}

\item{delay_incubation}{TODO add details}

\item{delay_onset_to_report}{TODO add details}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero? TODO to be described better.}

\item{...}{
  Arguments passed on to \code{\link[=convolve_delay_inputs]{convolve_delay_inputs}}, \code{\link[=.deconvolve_incidence_Richardson_Lucy]{.deconvolve_incidence_Richardson_Lucy}}
  \describe{
    \item{\code{n_report_time_steps}}{integer. Length of incidence time series.
Use only when providing empirical delay data.}
    \item{\code{delay_distribution}}{numeric square matrix or vector. TODO refactor to estimateR}
    \item{\code{threshold_chi_squared}}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}
    \item{\code{max_iterations}}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}
    \item{\code{verbose}}{Boolean. Print verbose output?}
  }}
}
\value{
module output object. Inferred incidence of infection events. TODO specify
}
\description{
This function reconstructs an incidence of infection events from incidence data representing delayed observations.
The assumption made is that delayed observations represent the convolution of the time series of infections with a delay distribution.
\code{deconvolve_incidence} implements a deconvolution algorithm (Richardson-Lucy) to reconstruct
a vector of infection events from input data representing delayed observations.
}
