% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve_incidence}
\alias{deconvolve_incidence}
\title{Infer Infection Events Dates from Delayed Observation}
\usage{
deconvolve_incidence(
  incidence_data,
  deconvolution_method = "Richardson-Lucy delay distribution",
  delay_incubation,
  delay_onset_to_report = c(1),
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution step).
Available options are:
\itemize{
\item{'Richardson-Lucy delay distribution',
implemented in \code{\link{.deconvolve_incidence_Richardson_Lucy}}}
}}

\item{delay_incubation}{TODO add details}

\item{delay_onset_to_report}{TODO add details}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero? TODO to be described better.}

\item{...}{
  Arguments passed on to \code{\link[=convolve_delay_inputs]{convolve_delay_inputs}}, \code{\link[=.deconvolve_incidence_Richardson_Lucy]{.deconvolve_incidence_Richardson_Lucy}}
  \describe{
    \item{\code{n_report_time_steps}}{integer. Length of incidence time series.
Use only when providing empirical delay data.}
    \item{\code{delay_distribution}}{numeric square matrix or vector. TODO refactor to estimateR}
    \item{\code{is_partially_reported_data}}{boolean value.
Set to \code{TRUE} if incidence was collected following a subsequent observation event.
For instance, if the incidence represents people starting to show symptoms of a disease
(dates of onset of symptoms), the data would typically have been collected among
individuals whose case was confirmed via a test.
If so, among all events of onset of symptoms, only those who had time to be
confirmed by a test were reported. Thus, close to the present, there
is an underreporting of onset of symptoms events.
In order to account for this effect, set this argument to \code{TRUE} and
input a \code{delay_distribution_final_report} argument.}
    \item{\code{delay_distribution_final_report}}{TODO refactor to estimateR
Distribution of the delay between the events collected in the incidence data
and the a posteriori observations of these events.
See \code{is_partially_reported_data} for more details.
In the example given there, \code{delay_distribution_final_report}
is the delay distribution between onset of symptoms and case confirmation.}
    \item{\code{threshold_chi_squared}}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}
    \item{\code{max_iterations}}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}
    \item{\code{verbose}}{Boolean. Print verbose output?}
  }}
}
\value{
module output object. Inferred incidence of infection events.
}
\description{
This function reconstructs an incidence of infection events from incidence data representing delayed observations.
The assumption made is that delayed observations represent the convolution of the infections with a delay distribution.
\code{deconvolve_incidence} implements a deconvolution algorithm (Richardson-Lucy) to reconstruct
a vector of infection events from input data representing delayed observations.
}
