% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve_incidence}
\alias{deconvolve_incidence}
\title{Infer infection events dates from delayed observations}
\usage{
deconvolve_incidence(
  incidence_data,
  deconvolution_method = "Richardson-Lucy delay distribution",
  delay_incubation,
  delay_onset_to_report = c(1),
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{An object containing incidence data through time.
It can either be:
\itemize{
\item A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}
\item A numeric vector. The vector corresponds to the \code{values} element
descrived above, and \code{index_offset} is implicitely zero.
This means that the first value in \code{incidence_data}
is associated with the reference time step (no shift towards the future or past).
}}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution step).
Available options are:
\itemize{
\item{'Richardson-Lucy delay distribution',
implemented in \code{\link{.deconvolve_incidence_Richardson_Lucy}}}
}}

\item{delay_incubation}{TODO add details}

\item{delay_onset_to_report}{TODO add details}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero?

@return A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the result of the computations on the input data.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the result is shifted compared to an \code{index_offset} of \code{0}.
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
Note that the \code{index_offset} of the output of the function call
accounts for the (optional) \code{index_offset} of the input.
}
If \code{index_offset} is \code{0} and \code{simplify_output = TRUE},
the \code{index_offset} is dropped and the \code{values}
element is returned as a numeric vector.}

\item{...}{
  Arguments passed on to \code{\link[=convolve_delay_inputs]{convolve_delay_inputs}}, \code{\link[=.deconvolve_incidence_Richardson_Lucy]{.deconvolve_incidence_Richardson_Lucy}}
  \describe{
    \item{\code{n_report_time_steps}}{integer. Length of incidence time series.
Use only when providing empirical delay data.}
    \item{\code{delay_distribution}}{numeric square matrix or vector.}
    \item{\code{threshold_chi_squared}}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}
    \item{\code{max_iterations}}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}
    \item{\code{verbose}}{Boolean. Print verbose output?}
  }}
}
\description{
This function reconstructs an incidence of infection events
from incidence data representing delayed observations.
The assumption made is that delayed observations represent
the convolution of the time series of infections with a delay distribution.
\code{deconvolve_incidence} implements a deconvolution algorithm (Richardson-Lucy) to reconstruct
a vector of infection events from input data that represents delayed observations.
}
