% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve_incidence}
\alias{deconvolve_incidence}
\title{Title}
\usage{
deconvolve_incidence(
  incidence_data,
  deconvolution_method = "Richardson-Lucy delay distribution",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
