% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve_incidence}
\alias{deconvolve_incidence}
\title{Infer Infection Events Dates from Delayed Observation}
\usage{
deconvolve_incidence(
  incidence_data,
  deconvolution_method = "Richardson-Lucy delay distribution",
  delay_incubation,
  delay_onset_to_report = c(1),
  start_date = NULL,
  time_step = "day",
  min_number_cases = NULL,
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric.}

\item{deconvolution_method}{string. Options are "Richardson-Lucy delay distribution"}

\item{min_number_cases}{}

\item{simplify_output}{boolean. Return a numeric vector instead of module output object if output offset is zero.}
}
\value{
module output object.
}
\description{
This function reconstructs an incidence of infection events from incidence data representing delayed observations.
The assumption made is that delayed observations represent the convolution of the infections with a delay distribution.
\code{deconvolve_incidence} implements a deconvolution algorithm (Richardson-Lucy) to reconstruct
a vector of infection events from input data representing delayed observations.
}
\examples{
#TODO add examples
}
