% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{delay_high}
\alias{delay_high}
\title{High-level delay parameters}
\arguments{
\item{delay_incubation}{TODO add details}

\item{delay_onset_to_report}{TODO add details}
}
\description{
High-level delay parameters
}
