% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{do_uncertainty_summary}
\alias{do_uncertainty_summary}
\title{Title}
\usage{
do_uncertainty_summary(
  original_values,
  bootstrapped_values,
  uncertainty_summary_method,
  value_col,
  bootstrap_id_col,
  index_col,
  output_Re_only,
  combine_bootstrap_and_estimation_uncertainties = FALSE,
  Re_HPDs = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
