% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.block_bootstrap_overlap_func}
\alias{.block_bootstrap_overlap_func}
\title{Helper function for block-bootstrapping}
\usage{
.block_bootstrap_overlap_func(
  incidence_vector,
  block_size = 10,
  keep_weekdays_aligned = TRUE
)
}
\arguments{
\item{incidence_vector}{numeric vector. Original incidence to bootstrap over.}

\item{block_size}{integer. Size of a bootstrapping block.}

\item{keep_weekdays_aligned}{boolean.
Set to \code{FALSE} if not daily incidence, or if no weekly noise pattern that would require
to apply errors to the same day of the week as they were in the original data.}
}
\value{
numeric vector. Bootstrapped differences.
}
\description{
Builds a bootstrapped vector of errors.
}
