% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_class_parameter_name}
\alias{.check_class_parameter_name}
\title{Utility functions for input validity.}
\usage{
.check_class_parameter_name(object, proper_class, parameter_name)
}
\arguments{
\item{object}{An object whose class needs checking,}

\item{proper_class}{A string describing the desired class of \code{object}.}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility function to check whether an object belongs to a particular class.
Wrapper function over \code{\link{.check_class}} needed because, being called from \code{\link{.are_valid_argument_values}},
the parameter name will not be the same as the one from the original function.
}
