% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.convolve_delay_distribution_vector_with_matrix}
\alias{.convolve_delay_distribution_vector_with_matrix}
\title{Title}
\usage{
.convolve_delay_distribution_vector_with_matrix(
  vector_a,
  matrix_b,
  vector_first = TRUE
)
}
\arguments{
\item{vector_first}{a boolean. Delay described in vector is applied before delay described in matrix}
}
\value{
square matrix. Delay distribution matrix
}
\description{
Title
}
