% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distribution_vectors}
\alias{.convolve_delay_distribution_vectors}
\title{Title}
\usage{
.convolve_delay_distribution_vectors(vector_a, vector_b)
}
\arguments{
\item{vector_a}{numeric}

\item{vector_b}{numeric}
}
\value{
numeric vector
}
\description{
Title
}
