% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distributions}
\alias{.convolve_delay_distributions}
\title{Convolve a list of delay vectors or matrices}
\usage{
.convolve_delay_distributions(delay_distributions)
}
\arguments{
\item{delay_distributions}{list of discretized delay distribution vector or matrix}
}
\value{
discretized delay distribution vector (if all input delays are
vectors) or matrix
}
\description{
Convolve a list of delay vectors or matrices
}
