% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convolution.R
\name{.convolve_delay_distributions}
\alias{.convolve_delay_distributions}
\title{Title}
\usage{
.convolve_delay_distributions(first_delay, second_delay)
}
\arguments{
\item{second_delay}{}
}
\value{
vector or matrix
}
\description{
Title
}
