% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{.deconvolve_incidence_Richardson_Lucy}
\alias{.deconvolve_incidence_Richardson_Lucy}
\title{Deconvolve the incidence input with the Richardson-Lucy (R-L) algorithm}
\usage{
.deconvolve_incidence_Richardson_Lucy(
  incidence_input,
  delay_distribution,
  threshold_chi_squared = 1,
  max_iterations = 100,
  verbose = FALSE
)
}
\arguments{
\item{incidence_input}{Module input object.
List with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}}

\item{delay_distribution}{numeric square matrix or vector. TODO refactor to estimateR}

\item{threshold_chi_squared}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}

\item{max_iterations}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}

\item{verbose}{Boolean. Print verbose output?}
}
\value{
module output object. Deconvolved incidence.
}
\description{
Deconvolve the incidence input with the Richardson-Lucy (R-L) algorithm
}
