% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{.deconvolve_incidence_Richardson_Lucy}
\alias{.deconvolve_incidence_Richardson_Lucy}
\title{Deconvolve the incidence input with the Richardson-Lucy (R-L) algorithm}
\usage{
.deconvolve_incidence_Richardson_Lucy(
  incidence_input,
  delay_distribution,
  threshold_chi_squared = 1,
  max_iterations = 100,
  verbose = FALSE
)
}
\arguments{
\item{incidence_input}{module input object.}

\item{delay_distribution}{numeric square matrix or vector. TODO refactor to estimateR}

\item{threshold_chi_squared}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}

\item{max_iterations}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}

\item{verbose}{Boolean. Print verbose output?}
}
\value{
module output object. Deconvolved incidence.
}
\description{
Deconvolve the incidence input with the Richardson-Lucy (R-L) algorithm
}
