% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{.deconvolve_incidence_Richardson_Lucy}
\alias{.deconvolve_incidence_Richardson_Lucy}
\title{Deconvolve the incidence input with the Richardson-Lucy (R-L) algorithm}
\usage{
.deconvolve_incidence_Richardson_Lucy(
  incidence_input,
  delay_distribution,
  is_partially_reported_data = FALSE,
  delay_distribution_final_report = NULL,
  threshold_chi_squared = 1,
  max_iterations = 100,
  verbose = FALSE
)
}
\arguments{
\item{incidence_input}{module input object.}

\item{delay_distribution}{numeric square matrix or vector. TODO refactor to estimateR}

\item{is_partially_reported_data}{boolean value.
Set to \code{TRUE} if incidence was collected following a subsequent observation event.
For instance, if the incidence represents people starting to show symptoms of a disease
(dates of onset of symptoms), the data would typically have been collected among
individuals whose case was confirmed via a test.
If so, among all events of onset of symptoms, only those who had time to be
confirmed by a test were reported. Thus, close to the present, there
is an underreporting of onset of symptoms events.
In order to account for this effect, set this argument to \code{TRUE} and
input a \code{delay_distribution_final_report} argument.}

\item{threshold_chi_squared}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}

\item{max_iterations}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}

\item{verbose}{Boolean. Print verbose output?}
}
\value{
module output object. Deconvolved incidence.
}
\description{
Deconvolve the incidence input with the Richardson-Lucy (R-L) algorithm
}
