% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Re.R
\name{.estimate_Re_EpiEstim_sliding_window}
\alias{.estimate_Re_EpiEstim_sliding_window}
\title{Estimate Re with EpiEstim using a sliding window}
\usage{
.estimate_Re_EpiEstim_sliding_window(
  incidence_input,
  minimul_cumul_incidence = 0,
  estimation_window = 3,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1
)
}
\arguments{
\item{incidence_input}{module input object. Time series of infections.}

\item{minimul_cumul_incidence}{Numeric scalar. Minimum number of cumulated infections before starting the Re estimation}

\item{estimation_window}{Integer scalar. Number of data points over which to assume Re to be constant.}

\item{mean_serial_interval}{Numeric positive scalar. "mean_si" for EpiEstim::estimate_R}

\item{std_serial_interval}{Numeric positive scalar. "std_si" for EpiEstim::estimate_R}

\item{mean_Re_prior}{Numeric positive scalar. "mean prior" for EpiEstim::estimate_R}
}
\value{
module output object. mean of Re estimates
}
\description{
The Re value reported for time t corresponds to the value estimated
when assuming that is Re is constant over e.g. (T-3, T-2, T-1, T),
for a sliding window of 4 time steps.
}
