% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.get_delay_matrix_from_delay_distribution_parms}
\alias{.get_delay_matrix_from_delay_distribution_parms}
\title{Build delay distribution matrix from list of delay distribution parameters}
\usage{
.get_delay_matrix_from_delay_distribution_parms(
  parm1_vector,
  parm2_vector,
  distribution_type = "gamma",
  max_quantile = 0.999
)
}
\arguments{
\item{max_quantile}{}
}
\value{
delay distribution matrix
}
\description{
Build delay distribution matrix from list of delay distribution parameters
}
