% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.get_delay_matrix_from_delay_distribution_parms}
\alias{.get_delay_matrix_from_delay_distribution_parms}
\title{Build delay distribution matrix from list of delay distributions}
\usage{
.get_delay_matrix_from_delay_distribution_parms(
  distributions,
  max_quantile = 0.999,
  offset_by_one = FALSE
)
}
\arguments{
\item{distributions}{list of distributions}

\item{offset_by_one}{A boolean}
}
\value{
delay distribution matrix
}
\description{
Build delay distribution matrix from list of delay distributions
}
