% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{.get_initial_deconvolution_shift}
\alias{.get_initial_deconvolution_shift}
\title{Get initial shift for deconvolution step}
\usage{
.get_initial_deconvolution_shift(delay_distribution_vector)
}
\arguments{
\item{delay_distribution_vector}{discretized probability distribution vector}
}
\value{
an integer value corresponding to the rounded median of
the input delay distribution.
}
\description{
This utility function returns the number of timesteps
by which the incidence data should be shifted back in the past
for the initial step of the Richardson-Lucy deconvolution algorithm.
}
