% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.get_matrix_from_single_delay_distr}
\alias{.get_matrix_from_single_delay_distr}
\title{Make delay distribution matrix from vector of delay distribution.}
\usage{
.get_matrix_from_single_delay_distr(delay_distribution_vector, N)
}
\arguments{
\item{delay_distribution_vector}{discretized probability distribution vector}

\item{N}{integer. Dimension of output matrix.}
}
\value{
discretized delay distribution matrix, representing a constant-through-time
delay distribution.
}
\description{
Make delay distribution matrix from vector of delay distribution.
}
