% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_module_input}
\alias{.get_module_input}
\title{Transform input data into a module input object}
\usage{
.get_module_input(data)
}
\arguments{
\item{data}{TODO specify what format data can take}
}
\value{
module input object.
List with a \code{values} and an \code{index_offset} element.
}
\description{
The input can be a list containing a \code{values} element
and an \code{index_offset} element, potentially among others.
It can also be a vector containing numeric values.
}
