% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_module_output}
\alias{.get_module_output}
\title{Transform a result output of a module into a module output}
\usage{
.get_module_output(results, input, offset = 0)
}
\arguments{
\item{results}{numeric vector containing output of module operations.}

\item{input}{module input object. Input originally given to module.}

\item{offset}{integer. Shift resulting from operations performed in module.}
}
\value{
module output object
}
\description{
Also takes the module input object to calculate the offset of the output object.
The new offset is simply (module input offset) + (shift applied during module operations)
The shift applied during module operations is the \code{offset} argument.
}
