% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.get_module_output}
\alias{.get_module_output}
\title{Transform a result output of a module into a module output 'object'}
\usage{
.get_module_output(results, input, offset = 0)
}
\arguments{
\item{results}{numeric vector containing output of module operations.}

\item{input}{Module input object.
List with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}

@return Module input object.
List with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}}

\item{offset}{integer. Shift resulting from operations performed in module.}
}
\description{
Also takes the module input object \code{input}
given to the module to calculate the offset of the output object.
The new offset is simply (module input offset) + (shift applied during module operations)
The shift applied during module operations is the \code{offset} argument.
}
