% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.get_module_output}
\alias{.get_module_output}
\title{Transform a result output of a module into a module output 'object'}
\usage{
.get_module_output(results, original_offset, additional_offset = 0)
}
\arguments{
\item{results}{numeric vector containing output of module operations.}

\item{original_offset}{integer. Input offset before computations.}

\item{additional_offset}{integer. Shift resulting from operations performed in module.}
}
\value{
A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the result of the computations on the input data.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the result is shifted compared to an \code{index_offset} of \code{0}.
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
Note that the \code{index_offset} of the output of the function call
accounts for the (optional) \code{index_offset} of the input.
}
If \code{index_offset} is \code{0} and \code{simplify_output = TRUE},
the \code{index_offset} is dropped and the \code{values}
element is returned as a numeric vector.
}
\description{
Also takes the module input object \code{input}
given to the module to calculate the offset of the output object.
The new offset is simply (module input offset) + (shift applied during module operations)
The shift applied during module operations is the \code{offset} argument.
}
