% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.get_right_boundary_for_distribution_vector}
\alias{.get_right_boundary_for_distribution_vector}
\title{Title}
\usage{
.get_right_boundary_for_distribution_vector(distribution, max_quantile)
}
\arguments{
\item{max_quantile}{}
}
\value{

}
\description{
Title
}
