% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-distribution.R
\name{.is_valid_distribution}
\alias{.is_valid_distribution}
\title{Title}
\usage{
.is_valid_distribution(distribution)
}
\arguments{
\item{distribution}{}
}
\value{
a boolean value.
}
\description{
Title
}
