% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simplify_output}
\alias{.simplify_output}
\title{Simplify output object if possible}
\usage{
.simplify_output(output)
}
\arguments{
\item{output}{Module output object}
}
\value{
numeric vector or module output object. simplified output
}
\description{
If offset is 0, return only vector containing values.
}
