% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{.smooth_LOESS}
\alias{.smooth_LOESS}
\title{LOESS smoothing function}
\usage{
.smooth_LOESS(incidence_input, data_points_incl = 21, degree = 1)
}
\arguments{
\item{incidence_input}{module input. Noisy incidence data to smooth.}

\item{data_points_incl}{integer. Size of the window used in the LOESS algorithm.}

\item{degree}{integer. LOESS degree.}
}
\value{
module output. Smoothed incidence.
}
\description{
see help page of \link{stats::loess} for more details.
}
\details{
#TODO add details on how data_points_incl relates to span.
}
