% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{.smooth_LOESS}
\alias{.smooth_LOESS}
\title{LOESS smoothing function}
\usage{
.smooth_LOESS(
  incidence_input,
  data_points_incl = 21,
  degree = 1,
  initial_Re_estimate_window = 5
)
}
\arguments{
\item{incidence_input}{Module input object.
List with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}

@return Module input object.
List with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}}

\item{data_points_incl}{integer. Size of the window used in the LOESS algorithm.
The \code{span} parameter passed to \code{\link[stats]{loess}} is computed as
the ratio of \code{data_points_incl} and the number of time steps in the input data.}

\item{degree}{integer. LOESS degree. Must be 0, 1 or 2.}

\item{initial_Re_estimate_window}{integer. In order to help with the smoothing, the function extends
the data back in time, padding with values obtained by assuming a constant Re. This parameter represents
the number of timesteps in the beginning of \code{incidence_input} to take into account when computing
the average initial Re.}
}
\description{
Prefer the use of the wrapper function \code{smooth_incidence(..., smoothing_method = "LOESS")}
instead of \code{.smooth_LOESS}.
}
\details{
This function implements the LOESS method for smoothing noisy data.
It relies on \code{\link[stats]{loess}}.
See the help section for \code{\link[stats]{loess}} for details on LOESS.
}
