% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{.summarise_bagged_mean}
\alias{.summarise_bagged_mean}
\title{Title}
\usage{
.summarise_bagged_mean(
  bootstrapped_estimates,
  include_original_estimates = TRUE,
  Re_estimate_col = "R_mean",
  bootstrap_id_col = "bootstrap_id",
  time_step = "day"
)
}
\arguments{
\item{bootstrap_id_col}{}
}
\value{

}
\description{
Title
}
