% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{.summarise_bagged_mean}
\alias{.summarise_bagged_mean}
\title{Compute bagged mean from bootstrapped replicates}
\usage{
.summarise_bagged_mean(
  bootstrapped_values,
  original_values = NULL,
  value_col,
  bootstrap_id_col,
  index_col
)
}
\arguments{
\item{bootstrapped_values}{Estimates obtained on a number of bootstrap
replicates of the original data.
Must be a dataframe in the long format
with a timestep index column named \code{index_col},
a bootstrap replicate index column named \code{bootstrap_id_col},
and an estimate column named \code{value_col}.
The index column must not contain any \code{NA} value.}

\item{original_values}{Optional. Values of reference
used to construct the uncertainty interval around.
Typically, these are estimates obtained on the original data.
Must be a dataframe with a timestep index column named \code{index_col}
and an estimate column named \code{value_col}.
The index column must not contain any \code{NA} value.}

\item{value_col}{string. Name of the column containing values.}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to the estimate on the original data.}

\item{index_col}{string. Name of the index column.
The index tracks which data point in bootstrapped estimates
corresponds to which data point in the original estimates.}
}
\value{
a dataframe containing a time step index column named \code{index_col}
and a column containing bagged mean values called \code{value_col}
}
\description{
If \code{original_values} are included,
these values are included in the mean computation
along with the \code{bootstrapped_values}.
}
