% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{.summarise_CI_bootstrap}
\alias{.summarise_CI_bootstrap}
\title{Build a confidence interval from bootstrapped values}
\usage{
.summarise_CI_bootstrap(
  central_values,
  bootstrapped_values,
  value_col,
  bootstrap_id_col,
  index_col,
  alpha = 0.95,
  prefix_up = "CI_up",
  prefix_down = "CI_down"
)
}
\arguments{
\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to the estimate on the original data.}

\item{index_col}{string. Index column to keep track of which data point
in bootstrapped estimates corresponds to which data point in the original estimates.}

\item{alpha}{value between 0 and 1. Confidence level of the confidence interval.}
}
\value{
dataframe with 4 columns:
\itemize{
\item{\code{index_col}, the timestep index column}
\item{\code{value_col}, the value input in \code{central_values}}
\item{CI_up, the upper limit of the confidence interval}
\item{CI_down, the lower limit of the confidence interval}
}
}
\description{
Build a confidence interval from bootstrapped values
}
