% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{.summarise_CI_bootstrap}
\alias{.summarise_CI_bootstrap}
\title{Build a confidence interval from bootstrapped values}
\usage{
.summarise_CI_bootstrap(
  central_values,
  bootstrapped_values,
  value_col,
  bootstrap_id_col,
  index_col,
  alpha = 0.95,
  prefix_up = "CI_up",
  prefix_down = "CI_down"
)
}
\arguments{
\item{central_values}{Values around which the confidence interval is going to be centered.
Must be a dataframe with a timestep index column named \code{index_col}
and a value column named \code{value_col}.
The index column must not contain any \code{NA} value.}

\item{bootstrapped_values}{Bootstrap
replicates of the original data.
Must be a dataframe in the long format
with a timestep index column named \code{index_col},
a bootstrap replicate index column named \code{bootstrap_id_col},
and a value column named \code{value_col}.
The index column must not contain any \code{NA} value.}

\item{value_col}{string. Name of the column containing values.}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to values associated to the original data.}

\item{index_col}{string. Name of the index column.
The index tracks which data point in bootstrapped values
corresponds to which data point in the original values.}

\item{alpha}{value between 0 and 1. Confidence level of the confidence interval.}

\item{prefix_up}{string. prefix to add to \code{output_value_col}
to name the column containing the upper limit of confidence interval}

\item{prefix_down}{string. prefix to add to \code{output_value_col}
to name the column containing the lower limit of confidence interval}
}
\value{
dataframe with 4 columns:
\itemize{
\item{\code{index_col}, the timestep index column}
\item{\code{value_col}, the value input in \code{central_values}}
\item{CI_up, the upper limit of the confidence interval}
\item{CI_down, the lower limit of the confidence interval}
}
}
\description{
Build a confidence interval from bootstrapped values
}
