% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{.summarise_CI_bootstrap}
\alias{.summarise_CI_bootstrap}
\title{Title}
\usage{
.summarise_CI_bootstrap(
  bootstrapped_estimates,
  alpha = 0.95,
  Re_estimate_col = "R_mean",
  bootstrap_id_col = "bootstrap_id",
  time_step = "day"
)
}
\arguments{
\item{time_step}{}
}
\value{

}
\description{
Title
}
