% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Re.R
\name{estimate_Re}
\alias{estimate_Re}
\title{Estimate effective reproductive number Re from incidence data}
\usage{
estimate_Re(
  incidence_data,
  estimation_method = "EpiEstim sliding window",
  simplify_output = FALSE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{estimation_method}{string. Method used to estimate reproductive number
values through time from the reconstructed infection timings.
Available options are:
\itemize{
\item{'EpiEstim sliding window',
implemented in \code{\link{.estimate_Re_EpiEstim_sliding_window}}}
\item{'EpiEstim piecewise constant',
implemented in \code{\link{.estimate_Re_EpiEstim_piecewise_constant}}}
}}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero? TODO to be described better.}

\item{...}{
  Arguments passed on to \code{\link[=.estimate_Re_EpiEstim_sliding_window]{.estimate_Re_EpiEstim_sliding_window}}
  \describe{
    \item{\code{minimum_cumul_incidence}}{Numeric scalar. Minimum number of cumulated infections before starting the Re estimation}
    \item{\code{estimation_window}}{Integer scalar. Number of data points over which to assume Re to be constant.}
    \item{\code{mean_serial_interval}}{Numeric positive scalar. \code{mean_si} for \code{\link[EpiEstim]{estimate_R}}}
    \item{\code{std_serial_interval}}{Numeric positive scalar. \code{std_si} for \code{\link[EpiEstim]{estimate_R}}}
    \item{\code{mean_Re_prior}}{Numeric positive scalar. \code{mean prior} for \code{\link[EpiEstim]{estimate_R}}}
  }}
}
\value{
a module output object. Re estimates.
}
\description{
The incidence data input should represent infections,
as opposed to representing delayed observations of infections.
If the incidence data represents delayed observations of infections
then one should deconvolve it first with \code{deconvolve_incidence}.
TODO add details on inner function, that it wraps around EpiEstim and so on.
}
