% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Re.R
\name{estimate_Re}
\alias{estimate_Re}
\title{Estimate effective reproductive number Re from incidence data}
\usage{
estimate_Re(
  incidence_data,
  estimation_method = "EpiEstim sliding window",
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{Format still to define.}

\item{estimation_method}{string. Options are "EpiEstim sliding window".}

\item{simplify_output}{boolean. Return a numeric vector instead of module output object if output offset is zero.}

\item{...}{#TODO figure out output format}
}
\value{
a module output object (subject to change)
}
\description{
The incidence data input should represent infections,
as opposed to representing delayed observations of infections.
If the incidence data represents delayed observations of infections
then one should deconvolve it first with \code{deconvolve_incidence}.
}
\details{
For now, only one Re estimation function is implemented.

#TODO specify input format
}
\examples{
#TODO add examples
}
