% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_block_bootstrapped_estimate}
\alias{get_block_bootstrapped_estimate}
\title{Estimate Re from incidence and estimate uncertainty with block-bootstrapping}
\usage{
get_block_bootstrapped_estimate(
  incidence_data,
  N_bootstrap_replicates = 100,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  delay_incubation,
  delay_onset_to_report = c(1),
  estimation_window = 3,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1,
  ref_date = NULL,
  time_step = "day",
  verbose = FALSE
)
}
\arguments{
\item{incidence_data}{numeric. TODO define acceptable format}

\item{N_bootstrap_replicates}{integer. Number of bootstrap samples.}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Options are: "LOESS", implemented in \code{\link{smooth_LOESS}}.}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution).
Options are "Richardson-Lucy delay distribution".}

\item{estimation_method}{string. Method used to estimate reproductive number
values through time from the reconstructed infection timings.
Options are "EpiEstim sliding window".}

\item{uncertainty_summary_method}{string. 'NONE' or one of the possible strings in \code{\link{summarise_uncertainty}}}

\item{estimation_window}{integer. Only use if \code{estimation_method = "EpiEstim sliding window"}.}

\item{mean_serial_interval}{numeric. see \code{\link{estimate_Re}}}

\item{std_serial_interval}{numeric. see \code{\link{estimate_Re}}}

\item{mean_Re_prior}{numeric value. Mean of prior distribution on Re.}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints. "day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)

#TODO add details on output formatting}

\item{verbose}{boolean. Print verbose output?}
}
\value{
tibble. Re estimations along with results from each pipeline step.
}
\description{
An estimation of the effective reproductive number through time is made with \code{smooth_deconvolve_estimate}
on the original incidence data.
Then, the same estimation is performed on a number of bootstrap samples built from the original incidence data.
The estimate on the original data is output along with confidence interval boundaries
built from the distribution of bootstrapped estimates.
}
\details{
A base assumption made here is that the delay between infection can be split into two independent delays
(with the second delay being optional).
The first waiting time corresponds to the time between the infection event and the onset of symptoms.
The second (optional) waiting time corresponds to the delay between the onset of symptoms and the case observation.
If this second waiting time is omitted, observation is assumed to have happened at the end of the first waiting time.
In latter case, the first waiting time distribution does not need to correspond to an incubation period per se,
but it must in any case correspond to the delay between infection and case observation.

#TODO clarify input
}
