% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_block_bootstrapped_estimate}
\alias{get_block_bootstrapped_estimate}
\title{Estimate Re from incidence and estimate uncertainty with block-bootstrapping}
\usage{
get_block_bootstrapped_estimate(
  incidence_data,
  N_bootstrap_replicates = 100,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  delay_incubation,
  delay_onset_to_report = c(1),
  ref_date = NULL,
  time_step = "day",
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{N_bootstrap_replicates}{integer. Number of bootstrap samples.}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution step).
Available options are:
\itemize{
\item{'Richardson-Lucy delay distribution',
implemented in \code{\link{.deconvolve_incidence_Richardson_Lucy}}}
}}

\item{estimation_method}{string. Method used to estimate reproductive number
values through time from the reconstructed infection timings.
Available options are:
\itemize{
\item{'EpiEstim sliding window',
implemented in \code{\link{.estimate_Re_EpiEstim_sliding_window}}}
}}

\item{uncertainty_summary_method}{string. One of the following options:
\itemize{
\item{'NONE' if no summary of bootstrap estimates is required}
\item{'original estimate - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the original estimates.}
\item{'bagged mean - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the mean of bootstrapped estimates and original estimates.}
}}

\item{delay_incubation}{TODO add details}

\item{delay_onset_to_report}{TODO add details}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{...}{
  Arguments passed on to \code{\link[=.smooth_LOESS]{.smooth_LOESS}}, \code{\link[=.deconvolve_incidence_Richardson_Lucy]{.deconvolve_incidence_Richardson_Lucy}}, \code{\link[=.estimate_Re_EpiEstim_sliding_window]{.estimate_Re_EpiEstim_sliding_window}}
  \describe{
    \item{\code{data_points_incl}}{integer. Size of the window used in the LOESS algorithm.}
    \item{\code{degree}}{integer. LOESS degree.}
    \item{\code{delay_distribution}}{numeric square matrix or vector. TODO refactor to estimateR}
    \item{\code{is_partially_reported_data}}{boolean value.
Set to \code{TRUE} if incidence was collected following a subsequent observation event.
For instance, if the incidence represents people starting to show symptoms of a disease
(dates of onset of symptoms), the data would typically have been collected among
individuals whose case was confirmed via a test.
If so, among all events of onset of symptoms, only those who had time to be
confirmed by a test were reported. Thus, close to the present, there
is an underreporting of onset of symptoms events.
In order to account for this effect, set this argument to \code{TRUE} and
input a \code{delay_distribution_final_report} argument.}
    \item{\code{delay_distribution_final_report}}{TODO refactor to estimateR
Distribution of the delay between the events collected in the incidence data
and the a posteriori observations of these events.
See \code{is_partially_reported_data} for more details.
In the example given there, \code{delay_distribution_final_report}
is the delay distribution between onset of symptoms and case confirmation.}
    \item{\code{threshold_chi_squared}}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}
    \item{\code{max_iterations}}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}
    \item{\code{verbose}}{Boolean. Print verbose output?}
    \item{\code{minimul_cumul_incidence}}{Numeric scalar. Minimum number of cumulated infections before starting the Re estimation}
    \item{\code{estimation_window}}{Integer scalar. Number of data points over which to assume Re to be constant.}
    \item{\code{mean_serial_interval}}{Numeric positive scalar. \code{mean_si} for \code{\link[EpiEstim]{estimate_R}}}
    \item{\code{std_serial_interval}}{Numeric positive scalar. \code{std_si} for \code{\link[EpiEstim]{estimate_R}}}
    \item{\code{mean_Re_prior}}{Numeric positive scalar. \code{mean prior} for \code{\link[EpiEstim]{estimate_R}}}
  }}
}
\value{
tibble. Re estimations along with results from each pipeline step. TODO add details
}
\description{
An estimation of the effective reproductive number through time is made with \code{smooth_deconvolve_estimate}
on the original incidence data.
Then, the same estimation is performed on a number of bootstrap samples built from the original incidence data.
The estimate on the original data is output along with confidence interval boundaries
built from the distribution of bootstrapped estimates.
}
\details{
A base assumption made here is that the delay between infection can be split into two independent delays
(with the second delay being optional).
The first waiting time corresponds to the time between the infection event and the onset of symptoms.
The second (optional) waiting time corresponds to the delay between the onset of symptoms and the case observation.
If this second waiting time is omitted, observation is assumed to have happened at the end of the first waiting time.
In latter case, the first waiting time distribution does not need to correspond to an incubation period per se,
but it must in any case correspond to the delay between infection and case observation.
}
