% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstrap_replicate}
\alias{get_bootstrap_replicate}
\title{Obtain a bootstrap replicate from incidence data}
\usage{
get_bootstrap_replicate(
  incidence_data,
  bootstrapping_method = "non-parametric block boostrap",
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{An object containing incidence data through time.
It can either be:
\itemize{
\item A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the incidence recorded on consecutive time steps.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the first value in \code{values} is shifted compared to a reference time step
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
}
\item A numeric vector. The vector corresponds to the \code{values} element
descrived above, and \code{index_offset} is implicitely zero.
This means that the first value in \code{incidence_data}
is associated with the reference time step (no shift towards the future or past).
}}

\item{bootstrapping_method}{string. Method to perform bootstrapping
of the original incidence data.
Available options are:
\itemize{
\item{'non-parametric block boostrap',
implemented in \code{\link{.block_bootstrap}}}
}}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero?}

\item{...}{
  Arguments passed on to \code{\link[=.block_bootstrap]{.block_bootstrap}}
  \describe{
    \item{\code{round_incidence}}{boolean. If \code{TRUE}, the bootstrapped incidence is rounded to the nearest integer.}
    \item{\code{smoothing_method}}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}
  }}
}
\value{
A list with two elements:
\enumerate{
\item A numeric vector named \code{values}: the result of the computations on the input data.
\item An integer named \code{index_offset}: the offset, counted in number of time steps,
by which the result is shifted compared to an \code{index_offset} of \code{0}.
This parameter allows one to keep track of the date of the first value in \code{values}
without needing to carry a \code{date} column around.
A positive offset means \code{values} are delayed in the future compared to the reference values.
A negative offset means the opposite.
Note that the \code{index_offset} of the output of the function call
accounts for the (optional) \code{index_offset} of the input.
}
If \code{index_offset} is \code{0} and \code{simplify_output = TRUE},
the \code{index_offset} is dropped and the \code{values}
element is returned as a numeric vector.
}
\description{
Apply a bootstrapping procedure on incidence data.
Estimating Re over many bootstrapped replicates allows one to estimate
the uncertainty over the estimated Re value due to observation noise.
For now, only a non-parametric block bootstrapping function is implemented.
}
