% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstrap_replicate}
\alias{get_bootstrap_replicate}
\title{Obtain a bootstrap replicate from incidence data}
\usage{
get_bootstrap_replicate(
  incidence_data,
  bootstrapping_method = "non-parametric block boostrap",
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{bootstrapping_method}{string. Method to perform bootstrapping
of the original incidence data.
Available options are:
\itemize{
\item{'non-parametric block boostrap',
implemented in \code{\link{.block_bootstrap}}}
}}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero? TODO to be described better.}

\item{...}{
  Arguments passed on to \code{\link[=.block_bootstrap]{.block_bootstrap}}
  \describe{
    \item{\code{round_incidence}}{boolean. If \code{TRUE}, the bootstrapped incidence is rounded to the nearest integer.}
    \item{\code{smoothing_method}}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}
  }}
}
\value{
a module output object. A boostrapped replicate. TODO improve
}
\description{
Apply a bootstrapping procedure on some original incidence data.
Estimating Re over many bootstrapped replicates allows one to estimate
the uncertainty over the estimated Re value due to observation noise.
In future updates, the user will be able to choose between different
bootstrapping procedures.
For now, only one bootstrapping function is implemented.
It performs a non-parametric block bootstrapping.
}
