% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstrap_replicate}
\alias{get_bootstrap_replicate}
\title{Obtain a bootstrap replicate from incidence data}
\usage{
get_bootstrap_replicate(
  original_data,
  bootstrapping_method = "non-parametric block boostrap",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Obtain a bootstrap replicate from incidence data
}
