% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_bootstrap_replicate}
\alias{get_bootstrap_replicate}
\title{Obtain a bootstrap replicate from incidence data}
\usage{
get_bootstrap_replicate(
  incidence_data,
  bootstrapping_method = "non-parametric block boostrap",
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{TODO Format still to define.}

\item{simplify_output}{boolean. Return a numeric vector instead of module output object if output offset is zero.}

\item{...}{Additional parameters. TODO add details}
}
\value{
a module output object. A boostrapped replicate.
}
\description{
Apply a bootstrapping procedure on some original incidence data.
Estimating Re over many bootstrapped replicates allows one to estimate
the uncertainty over the estimated Re value due to observation noise.
In future versions, the user will be able to choose between different
bootstrapping procedures.
For now, only one bootstrapping function is implemented.
It performs a non-parametric block bootstrapping.
}
