% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_bootstrapped_estimates_from_combined_observations}
\alias{get_bootstrapped_estimates_from_combined_observations}
\title{Estimate Re from incidence and estimate uncertainty by bootstrapping}
\usage{
get_bootstrapped_estimates_from_combined_observations(
  partially_delayed_incidence,
  fully_delayed_incidence,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  bootstrapping_method = "non-parametric block boostrap",
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  combine_bootstrap_and_estimation_uncertainties = FALSE,
  N_bootstrap_replicates = 100,
  delay_until_partial,
  delay_from_partial_to_full,
  partial_observation_requires_full_observation = TRUE,
  ref_date = NULL,
  time_step = "day",
  output_Re_only = TRUE,
  ...
)
}
\arguments{
\item{partially_delayed_incidence}{TODO add details}

\item{fully_delayed_incidence}{TODO add details}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution step).
Available options are:
\itemize{
\item{'Richardson-Lucy delay distribution',
implemented in \code{\link{.deconvolve_incidence_Richardson_Lucy}}}
}}

\item{estimation_method}{string. Method used to estimate reproductive number
values through time from the reconstructed infection timings.
Available options are:
\itemize{
\item{'EpiEstim sliding window',
implemented in \code{\link{.estimate_Re_EpiEstim_sliding_window}}}
\item{'EpiEstim piecewise constant',
implemented in \code{\link{.estimate_Re_EpiEstim_piecewise_constant}}}
}}

\item{bootstrapping_method}{string. Method to perform bootstrapping
of the original incidence data.
Available options are:
\itemize{
\item{'non-parametric block boostrap',
implemented in \code{\link{.block_bootstrap}}}
}}

\item{uncertainty_summary_method}{string. One of the following options:
\itemize{
\item{'NONE' if no summary of bootstrap estimates is required}
\item{'original estimate - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the original estimates.}
\item{'bagged mean - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the mean of bootstrapped estimates and original estimates.}
}}

\item{N_bootstrap_replicates}{integer. Number of bootstrap samples.}

\item{delay_until_partial}{TODO add details}

\item{delay_from_partial_to_full}{TODO add details}

\item{partial_observation_requires_full_observation}{boolean TODO add details}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{output_Re_only}{boolean. Should the output only contain Re estimates?
(as opposed to containing results for each intermediate step)}

\item{...}{
  Arguments passed on to \code{\link[=.smooth_LOESS]{.smooth_LOESS}}, \code{\link[=.deconvolve_incidence_Richardson_Lucy]{.deconvolve_incidence_Richardson_Lucy}}, \code{\link[=.estimate_Re_EpiEstim_sliding_window]{.estimate_Re_EpiEstim_sliding_window}}, \code{\link[=merge_outputs]{merge_outputs}}, \code{\link[=correct_for_partially_observed_data]{correct_for_partially_observed_data}}
  \describe{
    \item{\code{data_points_incl}}{integer. Size of the window used in the LOESS algorithm.}
    \item{\code{degree}}{integer. LOESS degree.}
    \item{\code{delay_distribution}}{numeric square matrix or vector. TODO refactor to estimateR}
    \item{\code{threshold_chi_squared}}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}
    \item{\code{max_iterations}}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}
    \item{\code{verbose}}{Boolean. Print verbose output?}
    \item{\code{estimation_window}}{Use with \code{estimation_method = "EpiEstim sliding window"}
Positive integer value.
Number of data points over which to assume Re to be constant.}
    \item{\code{minimum_cumul_incidence}}{Numeric value.
Minimum number of cumulated infections before starting the Re estimation.
Default is \code{12} as recommended in Cori et al., 2013.}
    \item{\code{mean_serial_interval}}{Numeric positive value. \code{mean_si} for \code{\link[EpiEstim]{estimate_R()}}}
    \item{\code{std_serial_interval}}{Numeric positive value. \code{std_si} for \code{\link[EpiEstim]{estimate_R()}}}
    \item{\code{mean_Re_prior}}{Numeric positive value. \code{mean prior} for \code{\link[EpiEstim]{estimate_R()}}}
    \item{\code{cutoff_observation_probability}}{value between 0 and 1.
Only datapoints for timesteps that have a probability to be observed higher
than \code{cutoff_observation_probability} are kept.
The few datapoints with a lower probability to be observed are trimmed off
the tail of the timeseries.}
  }}
}
\value{

}
\description{
Estimate Re from incidence and estimate uncertainty by bootstrapping
}
