% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_infections_from_incidence}
\alias{get_infections_from_incidence}
\title{Get timeseries of infection events from incidence data of delayed observations}
\usage{
get_infections_from_incidence(
  incidence_data,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  delay_incubation,
  delay_onset_to_report = c(1),
  is_partially_reported_data = FALSE,
  delay_distribution_final_report = NULL,
  output_infection_incidence_only = TRUE,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=merge_outputs]{merge_outputs}}, \code{\link[=correct_for_partially_observed_data]{correct_for_partially_observed_data}}
  \describe{
    \item{\code{ref_date}}{Date. Optional. Date of the first data entry in \code{incidence_data}}
    \item{\code{time_step}}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}
    \item{\code{cutoff_observation_probability}}{value between 0 and 1.
Only datapoints for timesteps that have a probability to be observed higher
than \code{cutoff_observation_probability} are kept.
The few datapoints with a lower probability to be observed are trimmed off
the tail of the timeseries.}
  }}
}
\value{

}
\description{
Get timeseries of infection events from incidence data of delayed observations
}
