% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-empirical_delays.R
\name{get_matrix_from_empirical_delay_distr}
\alias{get_matrix_from_empirical_delay_distr}
\title{Build matrix of delay distributions through time from empirical delay data.}
\usage{
get_matrix_from_empirical_delay_distr(
  empirical_delays,
  n_report_time_steps,
  ref_date = NULL,
  time_step = "day",
  min_number_cases = NULL,
  upper_quantile_threshold = 0.99,
  min_number_cases_fraction = 0.05,
  min_min_number_cases = 10,
  fit = "none",
  num_steps_in_a_unit = NULL
)
}
\arguments{
\item{empirical_delays}{dataframe containing the empirical data. See Details.}

\item{n_report_time_steps}{integer. Length of the incidence time series in the accompanying analysis.
This argument is needed to determine the dimensions of the output matrix.}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{min_number_cases}{integer.
Minimal number of cases to build the empirical distribution from.
If \code{num_steps_in_a_unit} is \code{NULL}, for any time step T,
the \code{min_number_cases} records prior to T are used.
If less than \code{min_number_cases} delays were recorded before T,
then T is ignored and the \code{min_number_cases} earliest-recorded delays are used.
If \code{num_steps_in_a_unit} is given a value, a similar same procedure is applied,
except that, now at least \code{min_number_cases} must be taken over a round number of
time units. For example, if \code{num_steps_in_a_unit = 7}, and time steps represent consecutive days,
to build the distribution for time step T,
we find the smallest number of weeks starting from T and going in the past,
for which at least \code{min_number_cases} delays were recorded.
We then use all the delays recorded during these weeks.
Weeks are not meant as necessarily being Monday to Sunday,
but simply 7 days in a row, e.g. it can be Thursday-Wednesday.
Again, if less than \code{min_number_cases} delays were recorded before T,
then T is ignored.
We then find the minimum number of weeks, starting from the first recorded delay
that contains at least \code{min_number_cases}.}

\item{upper_quantile_threshold}{numeric. Between 0 and 1.
Argument for internal use.}

\item{min_number_cases_fraction}{numeric. Between 0 and 1.
If \code{min_number_cases} is not provided (kept to \code{NULL}),
the number of most-recent cases used to build
the instant delay distribution is \code{min_number_cases_fraction}
times the total number of reported delays.}

\item{min_min_number_cases}{numeric. Lower bound
for number of cases used to build an instant delay distribution.}

\item{fit}{string. One of "gamma" or "none". Specifies the type of fit that
is applied to the columns of the delay matrix}

\item{num_steps_in_a_unit}{Optional argument.
Number of time steps in a full time unit (e.g. 7 if looking at weeks).
If set, the delays used to build a particular
delay distribution will span over a round number of such time units.
This option is included for comparison with legacy code.}
}
\value{
a discretized delay distribution matrix.
}
\description{
This function takes a record of delays between events and their observations
and builds a discretized delay distribution matrix from this record.
The discretized delay distribution matrix
is required for the application of the Richardson-Lucy algorithm.
The main benefit of providing empirical delay data to an \code{estimateR} analysis,
as opposed to specifiying a delay as a single distribution
(whether a fitted or empirical distribution) is that the variability of
the delays through time is used to inform the analysis and provide more accurate estimates.
If the average of delays has shifted from 5 days to 3 days between the beginning and end
of epidemic of interest, this will be reflected in the recorded empirical delays
and will be accounted for by \code{estimateR} when estimating the reproductive number.
}
\details{
The \code{ref_date} argument here will be understood as the date of the first record in the incidence data
for which the empirical delay data will be used.
If \code{ref_date} is not provided, the reference date will be taken as being the earliest date in the
\code{event_date} column of \code{empirical_delays}. In other words, the date of the first record in the incidence data
will be assumed to be the same as the date of the first record in the empirical delay data.
If this is not the case in your analysis, make sure to specify a \code{ref_date} argument.
}
