% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deconvolution.R
\name{get_matrix_from_empirical_delay_distr}
\alias{get_matrix_from_empirical_delay_distr}
\title{Build matrix of delay distributions through time from empirical delay data.}
\usage{
get_matrix_from_empirical_delay_distr(
  empirical_delays,
  n_report_time_steps,
  ref_date = NULL,
  time_step = "day",
  min_number_cases = NULL,
  upper_quantile_threshold = 0.99,
  min_number_cases_fraction = 0.05,
  min_min_number_cases = 10
)
}
\arguments{
\item{empirical_delays}{dataframe containing the empirical data. See Details.}

\item{n_report_time_steps}{integer. Length of the incidence time series in the accompanying analysis.
This argument is needed to determine the dimensions of the output matrix.}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{min_number_cases}{integer. Minimal number of cases to build
the empirical distribution from. TODO add details}

\item{upper_quantile_threshold}{numeric. Between 0 and 1. TODO add details}

\item{min_number_cases_fraction}{numeric. Between 0 and 1.
If \code{min_number_cases} is not provided (kept to \code{NULL}),
the number of most-recent cases used to build
the instant delay distribution is \code{min_number_cases_fraction}
times the total number of reported delays.}

\item{min_min_number_cases}{numeric. Lower bound
for number of cases used to build an instant delay distribution.}
}
\value{
a discretized delay distribution matrix.
}
\description{
This function takes a record of delays between events and their observations
and builds a discretized delay distribution matrix from this record.
The discretized delay distribution matrix
is required for the application of the Richardson-Lucy algorithm.
The main benefit of providing empirical delay data to an \code{estimateR} analysis,
as opposed to specifiying a delay as a single distribution
(whether a fitted or empirical distribution) is that the variability of
the delays through time is used to inform the analysis and provide more accurate estimates.
If the average of delays has shifted from 5 days to 3 days between the beginning and end
of epidemic of interest, this will be reflected in the recorded empirical delays
and will be accounted for by \code{estimateR} when estimating the reproductive number.
}
\details{
An \code{empirical_delays} dataframe must contain (at least) two columns.
An 'event_date' column of type \code{Date}
and a 'report_delay' column of type \code{numeric}.
Each row represents the recording of a single delay between event and observation.
Typically, the 'event' here is the onset of symptoms of the disease of interest.
And the observation can be, for instance, case confirmation, hospital admission,
admission to an ICU, or death, depending on what the incidence data represents.
For a particular row, 'event_date' would then represent, for a single individual,
the date at which symptoms appeared. And 'report_delay' would represent the number
of time steps (as specified by \code{time_step}) until the observation was made
for this same individual.
}
