% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{loss_function}
\alias{loss_function}
\title{Loss Function}
\usage{
loss_function(beta, X, y)
}
\arguments{
\item{beta}{A \code{vector} of coefficients associated with the features in xi.}

\item{X}{The \bold{ith} row of the predictors}

\item{y}{the \bold{ith} observation of the response}
}
\value{
`NULL`
}
\description{
Calculates the negative log likelihood p_i and plugs it into the loss function
}
\examples{
# Example usage:
beta <- c(1, 2)
X <- matrix(c(1, 1, 2, 3), ncol = 2)
y <- c(1, 0)
loss_function(beta, X, y)
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
