% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_delayed_observations}
\alias{simulate_delayed_observations}
\title{Simulate a series of delayed observations from a series of infections.}
\usage{
simulate_delayed_observations(
  infections,
  delay,
  noise = list(type = "noiseless")
)
}
\arguments{
\item{infections}{Positive integer vector. Course of infections through time.}

\item{delay}{Single delay or list of delays.
Each delay can be one of:
\itemize{
\item{a list representing a distribution object}
\item{a discretized delay distribution vector}
\item{a discretized delay distribution matrix}
\item{a dataframe containing empirical delay data}
}}

\item{noise}{List specifying the type of noise and its parameters, if applicable.}
}
\value{
Integer vector. Simulated delayed observations.
}
\description{
Simulate a series of delayed observations from a series of infections.
}
\examples{
## Basic usage of simulate_delayed_observations
# Simulating a series of delayed observations of infections generated by an 
# infection with a Re of 1.2. The delays of the observations follow a normal
# distribution.
set.seed(7)
infections <- simulate_infections(rep(1.5, 100))


delay <- list(name="norm", mean = 7, sd = 2)

delayed_observations_1 <- simulate_delayed_observations(
  infections,
  delay = delay
)

## Advanced usage of simulate_delayed_observations
# Simulating delayed observations using the same infections as above, but assuming
# the observation is delayed by a convolution of two different delays

shape_incubation = 3.2
scale_incubation = 1.3
delay_incubation <- list(name="gamma", shape = shape_incubation, scale = scale_incubation)

shape_onset_to_report = 2.7
scale_onset_to_report = 1.6
delay_onset_to_report <- list(name="gamma",
                              shape = shape_onset_to_report,
                              scale = scale_onset_to_report)

delayed_observations_2 <- simulate_delayed_observations(
  infections,
  delay = list(delay_incubation, delay_onset_to_report)
)

# Simulating noisy delayed observations, assuming a gaussian noise 
delayed_observations_3 <- simulate_delayed_observations(
  infections,
  delay = delay,
  noise = list(type = 'gaussian', sd = 0.8)
)
}
