% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_infections}
\alias{simulate_infections}
\title{Title}
\usage{
simulate_infections(Rt, imported_infections = 1, mean_SI = 4.8, sd_SI = 2.3)
}
\arguments{
\item{sd_SI}{}
}
\value{

}
\description{
Title
}
