% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{smooth_deconvolve_estimate}
\alias{smooth_deconvolve_estimate}
\title{Title}
\usage{
smooth_deconvolve_estimate(
  incidence_vector,
  delay_distribution_vector,
  initial_delta,
  minimul_cumul_incidence = 0,
  estimation_window = 3,
  mean_serial_interval = 4.8,
  std_serial_interval = 2.3,
  mean_Re_prior = 1,
  time_units_in_the_past = 30,
  max_iterations = 100,
  verbose = FALSE,
  output_Re_only = TRUE,
  ref_date = NULL,
  time_step = "day"
)
}
\arguments{
\item{time_step}{}
}
\value{

}
\description{
Title
}
