% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{smooth_deconvolve_estimate}
\alias{smooth_deconvolve_estimate}
\title{Estimate Re from incidence data}
\usage{
smooth_deconvolve_estimate(
  incidence_data,
  smoothing_method = "LOESS",
  deconvolution_method = "Richardson-Lucy delay distribution",
  estimation_method = "EpiEstim sliding window",
  delay_incubation,
  delay_onset_to_report = c(1),
  ref_date = NULL,
  time_step = "day",
  output_Re_only = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}

\item{deconvolution_method}{string. Method used to infer timings of infection
events from the original incidence data (aka deconvolution step).
Available options are:
\itemize{
\item{'Richardson-Lucy delay distribution',
implemented in \code{\link{.deconvolve_incidence_Richardson_Lucy}}}
}}

\item{estimation_method}{string. Method used to estimate reproductive number
values through time from the reconstructed infection timings.
Available options are:
\itemize{
\item{'EpiEstim sliding window',
implemented in \code{\link{.estimate_Re_EpiEstim_sliding_window}}}
}}

\item{delay_incubation}{TODO add details}

\item{delay_onset_to_report}{TODO add details}

\item{ref_date}{Date. Optional. Date of the first data entry in \code{incidence_data}}

\item{time_step}{string. Time between two consecutive incidence datapoints.
"day", "2 days", "week", "year"... (see \code{\link[base]{seq.Date}} for details)}

\item{output_Re_only}{boolean. Should the output only contain Re estimates?
(as opposed to containing results for each intermediate step)}

\item{...}{
  Arguments passed on to \code{\link[=.smooth_LOESS]{.smooth_LOESS}}, \code{\link[=.deconvolve_incidence_Richardson_Lucy]{.deconvolve_incidence_Richardson_Lucy}}, \code{\link[=.estimate_Re_EpiEstim_sliding_window]{.estimate_Re_EpiEstim_sliding_window}}
  \describe{
    \item{\code{data_points_incl}}{integer. Size of the window used in the LOESS algorithm.}
    \item{\code{degree}}{integer. LOESS degree.}
    \item{\code{delay_distribution}}{numeric square matrix or vector. TODO refactor to estimateR}
    \item{\code{threshold_chi_squared}}{numeric scalar. Threshold for chi-squared values under which the R-L algorithm stops.}
    \item{\code{max_iterations}}{integer. Maximum threshold for the number of iterations in the R-L algorithm.}
    \item{\code{verbose}}{Boolean. Print verbose output?}
    \item{\code{minimul_cumul_incidence}}{Numeric scalar. Minimum number of cumulated infections before starting the Re estimation}
    \item{\code{estimation_window}}{Integer scalar. Number of data points over which to assume Re to be constant.}
    \item{\code{mean_serial_interval}}{Numeric positive scalar. \code{mean_si} for \code{\link[EpiEstim]{estimate_R}}}
    \item{\code{std_serial_interval}}{Numeric positive scalar. \code{std_si} for \code{\link[EpiEstim]{estimate_R}}}
    \item{\code{mean_Re_prior}}{Numeric positive scalar. \code{mean prior} for \code{\link[EpiEstim]{estimate_R}}}
  }}
}
\value{
effective reproductive number estimates through time TODO add details
}
\description{
This pipe function combines a smoothing step using (to remove noise from the original observations),
a deconvolution step (to retrieve infection events from the observed delays),
and an Re estimation step wrapping around \code{\link[EpiEstim]{estimate_R}}.
}
\details{
The \code{\link[=smooth_incidence]{smoothing step}} uses the LOESS method by default.
The \code{\link[=deconvolve_incidence]{deconvolution step}} uses the Richardson-Lucy algorithm  by default.
The \code{\link[=estimate_Re]{Re estimation}} uses the Cori method with a sliding window  by default.
}
