% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_incidence}
\alias{smooth_incidence}
\title{Smooth noisy incidence data}
\usage{
smooth_incidence(
  incidence_data,
  smoothing_method = "LOESS",
  simplify_output = TRUE,
  ...
)
}
\arguments{
\item{incidence_data}{numeric vector. Incidence data (TODO improve definition)}

\item{smoothing_method}{string. Method used to smooth the original incidence data.
Available options are:
\itemize{
\item{'LOESS', implemented in \code{\link{.smooth_LOESS}}}
}}

\item{simplify_output}{boolean. Return a numeric vector instead of module
output object if output offset is zero? TODO to be described better.}

\item{...}{
  Arguments passed on to \code{\link[=.smooth_LOESS]{.smooth_LOESS}}
  \describe{
    \item{\code{data_points_incl}}{integer. Size of the window used in the LOESS algorithm.}
    \item{\code{degree}}{integer. LOESS degree.}
  }}
}
\value{
module output. Smoothed incidence. TODO add details
}
\description{
Currently only LOESS smoothing (\code{smoothing_method = "LOESS"}) is implemented.
}
