% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{summarise_uncertainty}
\alias{summarise_uncertainty}
\title{Summarise the uncertainty obtained from bootstrapping}
\usage{
summarise_uncertainty(
  bootstrapped_estimates,
  original_estimates = NULL,
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  Re_estimate_col = "R_mean",
  bootstrap_id_col = "bootstrap_id",
  index_col = "idx",
  ...
)
}
\arguments{
\item{bootstrapped_estimates}{TODO add details}

\item{original_estimates}{Optional. TODO add details}

\item{uncertainty_summary_method}{One of these options:
\itemize{
\item{'original estimate - CI from bootstrap estimates'}
\item{'bagged mean - CI from bootstrap estimates'}
}}

\item{Re_estimate_col}{string. Name of the column containing Re estimates}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to the estimate on the original data.}

\item{index_col}{string. Index column to keep track of which data point
in bootstrapped estimates corresponds to which data point in the original estimates.}

\item{...}{
  Arguments passed on to \code{\link[=.summarise_CI_bootstrap]{.summarise_CI_bootstrap}}
  \describe{
    \item{\code{alpha}}{value between 0 and 1. Confidence level.}
  }}
}
\value{
dataframe containing Re estimates (column 'Re_estimate')
and confidence interval boundaries.
}
\description{
Summarise the uncertainty obtained from bootstrapping
}
