% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{summarise_uncertainty}
\alias{summarise_uncertainty}
\title{Title}
\usage{
summarise_uncertainty(
  bootstrapped_estimates,
  original_estimates = NA,
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  Re_estimate_col = "R_mean",
  bootstrap_id_col = "bootstrap_id",
  time_step = "day"
)
}
\arguments{
\item{uncertainty_summary_method}{One of these strings: \itemize{
\item{'original estimate - CI from bootstrap estimates'}
\item{'bagged mean - CI from bootstrap estimates'}
}}

\item{Re_estimate_col}{string. Name of the column containing Re estimates}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to the estimate on the original data.}

\item{time_step}{string}
}
\value{
dataframe containing Re estimates and confidence interval boundaries.
}
\description{
Title
}
