% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{summarise_uncertainty}
\alias{summarise_uncertainty}
\title{Summarise the uncertainty obtained from bootstrapping}
\usage{
summarise_uncertainty(
  bootstrapped_values,
  original_values = NULL,
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  value_col = "Re_estimate",
  output_value_col = "Re_estimate",
  bootstrap_id_col = "bootstrap_id",
  index_col = "idx",
  ...
)
}
\arguments{
\item{uncertainty_summary_method}{One of these options:
\itemize{
\item{'original estimate - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the original estimates.}
\item{'bagged mean - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the mean of bootstrapped estimates and original estimates.}
}}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to the estimate on the original data.}

\item{index_col}{string. Index column to keep track of which data point
in bootstrapped estimates corresponds to which data point in the original estimates.}

\item{...}{
  Arguments passed on to \code{\link[=.summarise_CI_bootstrap]{.summarise_CI_bootstrap}}
  \describe{
    \item{\code{alpha}}{value between 0 and 1. Confidence level of the confidence interval.}
  }}
}
\value{
dataframe containing Re estimates (column 'Re_estimate')
and confidence interval boundaries, with 4 columns like so:
\itemize{
\item{\code{index_col}, the timestep index column}
\item{\code{Re_estimate_col}, Re estimates}
\item{CI_up, the upper limit of the confidence interval}
\item{CI_down, the lower limit of the confidence interval}
}
}
\description{
Summarise the uncertainty obtained from bootstrapping
}
