% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-uncertainty.R
\name{summarise_uncertainty}
\alias{summarise_uncertainty}
\title{Summarise the uncertainty obtained from bootstrapping}
\usage{
summarise_uncertainty(
  bootstrapped_values,
  original_values = NULL,
  uncertainty_summary_method = "original estimate - CI from bootstrap estimates",
  value_col = "Re_estimate",
  output_value_col = "Re_estimate",
  bootstrap_id_col = "bootstrap_id",
  index_col = "idx",
  ...
)
}
\arguments{
\item{bootstrapped_values}{Bootstrap
replicates of the original data.
Must be a dataframe in the long format
with a timestep index column named \code{index_col},
a bootstrap replicate index column named \code{bootstrap_id_col},
and a value column named \code{value_col}.
The index column must not contain any \code{NA} value.}

\item{original_values}{Optional. Values of reference
used to construct the uncertainty interval around.
Typically, these are estimates obtained on the original data.
Must be a dataframe with a timestep index column named \code{index_col}
and a value column named \code{value_col}.
The index column must not contain any \code{NA} value.}

\item{uncertainty_summary_method}{One of these options:
\itemize{
\item{'original estimate - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the original estimates.}
\item{'bagged mean - CI from bootstrap estimates'.
The confidence interval is built using bootstrapped estimates
and centered around the mean of bootstrapped estimates and original estimates.}
}}

\item{value_col}{string. Name of the column containing values.}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to values associated to the original data.}

\item{index_col}{string. Name of the index column.
The index tracks which data point in bootstrapped values
corresponds to which data point in the original values.}

\item{...}{
  Arguments passed on to \code{\link[=.summarise_CI_bootstrap]{.summarise_CI_bootstrap}}
  \describe{
    \item{\code{central_values}}{Values around which the confidence interval is going to be centered.
Must be a dataframe with a timestep index column named \code{index_col}
and a value column named \code{value_col}.
The index column must not contain any \code{NA} value.}
    \item{\code{alpha}}{value between 0 and 1. Confidence level of the confidence interval.}
  }}
}
\value{
A dataframe containing Re estimates (column 'Re_estimate')
and confidence interval boundaries, with 4 columns like so:
\itemize{
\item{\code{index_col}, the timestep index column}
\item{A column named \code{output_value_col},
containing the central values (typically these are Re estimates) }
\item{\code{CI_up}, upper limit of the confidence interval}
\item{\code{CI_down}, the lower limit of the confidence interval}
}
}
\description{
This function takes as input bootstrapped values: Re estimates or others
and builds uncertainty intervals from these bootstrapped values.
}
\details{
This function is not meant to be needed by regular users.
It is exported as it can be useful when building 'pipes':
sequences of computations that weave together the different
modules provided by \code{estimateR}.
}
