% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{uncertainty}
\alias{uncertainty}
\title{Uncertainty summary}
\arguments{
\item{original_estimates}{Optional. TODO add details}

\item{bootstrapped_estimates}{TODO add details}

\item{Re_estimate_col}{string. Name of the column containing Re estimates}

\item{bootstrap_id_col}{string. Name of the column containing bootstrap samples numbering.
Id 0 must correspond to the estimate on the original data.}

\item{index_col}{string. Index column to keep track of which data point
in bootstrapped estimates corresponds to which data point in the original estimates.}

\item{alpha}{value between 0 and 1. Confidence level.}
}
\description{
Uncertainty summary
}
