% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{validation_utility_params}
\alias{validation_utility_params}
\title{Utility functions for input validity.}
\arguments{
\item{string_user_input}{A string containing the value that the user passed for the tested string type parameter.}

\item{parameter_name}{A string containing the name the tested parameter had in the initial function in which it was passed.}

\item{incidence_data_length}{A number representing the length of the given incidence data.}

\item{user_inputs}{A list of lists with two elements: the first is the value of the parameter to be tested. The second is the expected type of that parameter.}

\item{number}{The value to be tested}

\item{user_input}{The variable to be tested}
}
\value{
TRUE if all tests were passed. Throws an error otherwise.
}
\description{
Utility functions for input validity.
}
